@@ -1,41 +1,27 @@
 {
   "name": "new-tab-pro-react",
-  "version": "0.1.0",
+  "version": "5.0.0",
   "private": true,
   "dependencies": {
-    "@types/gsap": "^1.20.2",
-    "@types/node": "^12.12.18",
-    "@types/react": "^16.8.23",
-    "@types/react-dom": "^16.8.5",
-    "@types/react-redux": "^7.1.1",
-    "@types/styled-components": "^4.4.1",
-    "@types/three": "^0.93.31",
-    "@types/webspeechapi": "0.0.29",
-    "gsap": "^2.1.3",
-    "lottie-web": "^5.5.7",
-    "mobx": "^5.15.1",
-    "mobx-react": "^6.1.4",
-    "node-sass": "^4.12.0",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-redux": "^6.0.1",
-    "react-scripts": "2.1.5",
-    "redux": "^4.0.4",
-    "styled-components": "^4.4.1",
-    "three": "^0.101.1",
-    "three-orbit-controls": "^82.1.0",
-    "typesafe-actions": "^3.4.0",
-    "typescript": "^3.5.3"
+    "@floating-ui/react": "^0.15.1",
+    "framer-motion": "^8.2.4",
+    "lottie-web": "^5.10.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-query": "^3.39.2",
+    "react-scripts": "^5.0.1",
+    "react-viewport-list": "^6.3.0",
+    "styled-components": "^5.3.6",
+    "three": "^0.147.0",
+    "zustand": "^4.2.0"
   },
   "scripts": {
     "start": "react-scripts start",
     "dev": "react-scripts start",
     "build": "react-scripts build",
+    "lint": "eslint ./src",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": "react-app"
   },
   "browserslist": [
     ">0.2%",
@@ -44,7 +30,18 @@
     "not op_mini all"
   ],
   "devDependencies": {
-    "@types/chrome": "0.0.79",
-    "raw-loader": "^4.0.0"
+    "@types/chrome": "^0.0.206",
+    "@types/node": "^18.11.12",
+    "@types/react-dom": "^18.0.9",
+    "@types/styled-components": "^5.1.26",
+    "@types/three": "^0.146.0",
+    "@types/webspeechapi": "^0.0.29",
+    "@typescript-eslint/eslint-plugin": "^5.46.0",
+    "@typescript-eslint/parser": "^5.46.0",
+    "eslint": "^8.29.0",
+    "eslint-plugin-react": "^7.31.11",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "raw-loader": "^4.0.2",
+    "typescript": "^4.9.4"
   }
 }
