@@ -1,8 +1,11 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import { App } from "./components/App";
-import options from "./store/options";
 
-options.load(() => {
-  ReactDOM.render(<App />, document.getElementById("root"));
-});
+const container = document.getElementById("root");
+
+if (container !== null) {
+  const root = createRoot(container);
+
+  root.render(<App />);
+}
