@@ -510,5 +510,5 @@ export default [
   [0.6015625, 0.41015625, 0.91796875],
   [0.65625, 0.4296875, 0.92578125],
   [0.703125, 0.453125, 0.9296875],
-  [0.7421875, 0.46875, 0.93359375]
+  [0.7421875, 0.46875, 0.93359375],
 ];
