@@ -1,6 +1,4 @@
-import React, { useState } from "react";
-import { reaction } from "mobx";
-import { observer } from "mobx-react";
+import React, { useCallback, useEffect, useState } from "react";
 import {
   Form,
   FormButtonsGroup,
@@ -10,98 +8,107 @@ import {
   FormFooter,
   FormTextField,
 } from "../../lib/form";
-import sites, { ISite } from "../../store/sites";
-import editSite from "../../store/edit-site";
+import { useSiteEditor, useSites } from "./store";
 
-export const EditSiteForm = observer(() => {
-  const [site, setSite]: [ISite, React.Dispatch<ISite>] = useState({
-    name: "",
-    url: "",
-    image: "",
-  });
+export const EditSiteForm = () => {
+  const [site, setSite] = useState<ISite>({ name: "", url: "", image: "" });
+  const getSite = useSites((state) => state.get);
+  const update = useSites((state) => state.update);
+  const index = useSiteEditor((state) => state.index);
+  const setIndex = useSiteEditor((state) => state.setIndex);
+  const active = useSiteEditor((state) => state.active);
+  const setActive = useSiteEditor((state) => state.setActive);
 
-  reaction(
-    () => editSite.id,
-    id => {
-      const site = sites.get(id);
-
-      if (site) {
-        setSite(site);
-      }
+  const onSiteNameChange = useCallback(
+    (event: React.ChangeEvent<HTMLInputElement>) => {
+      setSite((state) => ({ ...state, name: event.target.value }));
     },
+    [],
   );
 
-  const undo = () => {
-    const site = sites.get(editSite.id);
+  const onSiteUrlChange = useCallback(
+    (event: React.ChangeEvent<HTMLInputElement>) => {
+      setSite((state) => ({ ...state, url: event.target.value }));
+    },
+    [],
+  );
 
-    if (site) {
-      setSite(site);
+  const onSiteImageChange = useCallback(
+    (event: React.ChangeEvent<HTMLInputElement>) => {
+      setSite((state) => ({ ...state, image: event.target.value }));
+    },
+    [],
+  );
+
+  const undo = useCallback(() => {
+    if (index !== null) {
+      const site = getSite(index);
+
+      if (site !== undefined) {
+        setSite({ ...site });
+      }
     }
-  };
+  }, [getSite, index]);
+
+  const onClose = useCallback(() => {
+    setActive(false);
+    undo();
+  }, [setActive, undo]);
+
+  const onFormSubmit = useCallback((event: React.FormEvent) => {
+    event.preventDefault();
+  }, []);
+
+  const onUpdateClick = useCallback(
+    (event: React.MouseEvent) => {
+      event.preventDefault();
+      if (index !== null) {
+        update(index, { ...site });
+      }
+      setIndex(null);
+      setActive(false);
+    },
+    [index, setActive, setIndex, site, update],
+  );
+
+  useEffect(() => {
+    undo();
+  }, [index, undo]);
 
   return (
     <Form
-      active={editSite.active}
-      closeAction={() => {
-        editSite.active = false;
-      }}
-      onSubmit={e => e.preventDefault()}
+      active={active === true}
+      closeAction={onClose}
+      onSubmit={onFormSubmit}
     >
       <FormHeader>Edit Site</FormHeader>
       <FormBody>
         <FormTextField
           label="Site name"
           value={site.name}
-          onChange={e => {
-            setSite({
-              ...site,
-              name: e.target.value,
-            });
-          }}
+          onChange={onSiteNameChange}
         />
         <FormTextField
           label="Site url"
           value={site.url}
-          onChange={e => {
-            setSite({
-              ...site,
-              url: e.target.value,
-            });
-          }}
+          onChange={onSiteUrlChange}
         />
         <FormTextField
           label="Image url or data:base64"
           value={site.image}
-          onChange={e => {
-            setSite({
-              ...site,
-              image: e.target.value,
-            });
-          }}
+          onChange={onSiteImageChange}
         />
       </FormBody>
       <FormFooter>
         <FormButtonsGroup>
-          <FormButton
-            success
-            onClick={event => {
-              event.preventDefault();
-              if (editSite.id !== null) {
-                sites.update(editSite.id, {
-                  ...site,
-                });
-
-                editSite.active = false;
-              }
-            }}
-          >
+          <FormButton success onClick={onUpdateClick}>
             Update
           </FormButton>
-          <FormButton warn onClick={() => undo()}>
+          <FormButton warn onClick={undo}>
             Undo
           </FormButton>
         </FormButtonsGroup>
       </FormFooter>
     </Form>
   );
-});
+};
