@@ -1,13 +1,11 @@
-import React from "react";
-import { observer } from "mobx-react";
-import options from "./../../store/options";
-import sites from "../../store/sites";
-import editSite from "../../store/edit-site";
+import React, { useCallback } from "react";
+import { useControls, useOptions } from "./../../store/options";
 import { EditSiteForm } from "./EditSiteForm";
 import styled from "styled-components/macro";
 import { theme } from "../../theme/theme-default";
 import { AddSiteForm } from "./AddSiteForm";
 import { AnimationPanel } from "../animation-panel/AnimationPanel";
+import { useSiteEditor, useSites } from "./store";
 
 const SitesList = styled(AnimationPanel)`
   position: relative;
@@ -183,42 +181,71 @@ const SiteIconItem = styled(SiteItem)`
   }
 `;
 
-export const Sites = observer(() => {
-  if (options.showWebSites === false) {
+const Site = styled<
+  Styled<{
+    site: ISite;
+    index: number;
+    onEdit: (index: number) => void;
+    onRemove: (index: number) => void;
+  }>
+>(({ site, index, onEdit, onRemove }) => {
+  const additionalOptions = useOptions((state) => state.controls);
+
+  const onEditClick = useCallback(() => {
+    onEdit(index);
+  }, [index, onEdit]);
+
+  const onRemoveClick = useCallback(() => {
+    onRemove(index);
+  }, [index, onRemove]);
+
+  return (
+    <SiteItem>
+      <SiteHeader>
+        <SiteName>{site.name}</SiteName>
+        <EditSiteButton active={additionalOptions} onClick={onEditClick} />
+        <RemoveSiteButton active={additionalOptions} onClick={onRemoveClick} />
+      </SiteHeader>
+      <SiteLink href={site.url}>
+        <SiteImage src={site.image} />
+      </SiteLink>
+    </SiteItem>
+  );
+})``;
+
+export const Sites = () => {
+  const showWebSites = useOptions((state) => state.sites);
+  const openAddSite = useControls((state) => state.openAddSite);
+  const sites = useSites((state) => state.sites);
+  const remove = useSites((state) => state.remove);
+  const setIndex = useSiteEditor((state) => state.setIndex);
+  const setActive = useSiteEditor((state) => state.setActive);
+
+  const onEdit = useCallback(
+    (index: number) => {
+      setIndex(index);
+      setActive(true);
+    },
+    [setActive, setIndex],
+  );
+
+  if (showWebSites === false) {
     return null;
   }
 
   return (
     <React.Fragment>
       <SitesList>
-        {sites.sites.map((site, id) => (
-          <SiteItem key={`site-item-${id}`}>
-            <SiteHeader>
-              <SiteName>{site.name}</SiteName>
-              <EditSiteButton
-                active={options.additionalOptions}
-                onClick={() => {
-                  editSite.id = id;
-                  editSite.active = true;
-                }}
-              />
-              <RemoveSiteButton
-                active={options.additionalOptions}
-                onClick={() => {
-                  sites.remove(id);
-                }}
-              />
-            </SiteHeader>
-            <SiteLink href={site.url}>
-              <SiteImage src={site.image} />
-            </SiteLink>
-          </SiteItem>
+        {sites.map((site, index) => (
+          <Site
+            key={`site-item-${index}`}
+            site={site}
+            index={index}
+            onEdit={onEdit}
+            onRemove={remove}
+          />
         ))}
-        <SiteIconItem
-          onClick={() => {
-            options.showAddSiteForm = true;
-          }}
-        >
+        <SiteIconItem onClick={openAddSite}>
           <SiteIcon viewBox="0 0 301 301">
             <path d="M112.5.5h76c9,0,12,5,12,12v88h88c7.45,0,12,4.49,12,12v76c0,7.38-4.48,12-12,12h-88v88c0,8-5.5,12-16,12h-76c-7.62,0-8-5.75-8-12v-88h-88c-7.54,0-12-4-12-12v-76c0-7.73,5.23-12,12-12h88v-88C100.5,4.69,103.78.5,112.5.5Z" />
           </SiteIcon>
@@ -228,4 +255,4 @@ export const Sites = observer(() => {
       <AddSiteForm />
     </React.Fragment>
   );
-});
+};
