@@ -23,53 +23,53 @@ export const cloudyDay = {
             {
               "i": {
                 "x": [0.833],
-                "y": [0.833]
+                "y": [0.833],
               },
               "o": {
                 "x": [0.167],
-                "y": [0.167]
+                "y": [0.167],
               },
               "n": ["0p833_0p833_0p167_0p167"],
               "t": 1,
               "s": [0],
-              "e": [100]
+              "e": [100],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 0,
           "k": [
             -1.548,
             -127.143,
-            0
+            0,
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             7.917,
             12,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             3125.714,
             3125.714,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "shapes": [
@@ -86,100 +86,100 @@ export const cloudyDay = {
                   "i": [
                     [
                       0,
-                      6.403
+                      6.403,
                     ],
                     [
                       6.627,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      -6.627
+                      -6.627,
                     ],
                     [
                       -0.088,
-                      -0.587
+                      -0.587,
                     ],
                     [
                       0,
-                      -5.033
+                      -5.033,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0.011
-                    ]
+                      0.011,
+                    ],
                   ],
                   "o": [
                     [
                       0,
-                      -6.627
+                      -6.627,
                     ],
                     [
                       -6.627,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0.612
+                      0.612,
                     ],
                     [
                       -4.746,
-                      0.972
+                      0.972,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      -0.011
+                      -0.011,
                     ],
                     [
                       6.316,
-                      -0.347
-                    ]
+                      -0.347,
+                    ],
                   ],
                   "v": [
                     [
                       24,
-                      12
+                      12,
                     ],
                     [
                       12,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      12
+                      12,
                     ],
                     [
                       0.149,
-                      13.795
+                      13.795,
                     ],
                     [
                       -8.167,
-                      24
+                      24,
                     ],
                     [
                       12.667,
-                      24
+                      24,
                     ],
                     [
                       12.665,
-                      23.966
-                    ]
+                      23.966,
+                    ],
                   ],
-                  "c": true
+                  "c": true,
                 },
-                "ix": 2
+                "ix": 2,
               },
               "nm": "Path 1",
               "mn": "ADBE Vector Shape - Group",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "fl",
@@ -189,19 +189,19 @@ export const cloudyDay = {
                   0.874509811401,
                   0.878431379795,
                   0.89411765337,
-                  1
+                  1,
                 ],
-                "ix": 4
+                "ix": 4,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 5
+                "ix": 5,
               },
               "r": 1,
               "nm": "Fill 1",
               "mn": "ADBE Vector Graphic - Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "tr",
@@ -209,61 +209,61 @@ export const cloudyDay = {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "a": {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 1
+                "ix": 1,
               },
               "s": {
                 "a": 0,
                 "k": [
                   100,
-                  100
+                  100,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "r": {
                 "a": 0,
                 "k": 0,
-                "ix": 6
+                "ix": 6,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 7
+                "ix": 7,
               },
               "sk": {
                 "a": 0,
                 "k": 0,
-                "ix": 4
+                "ix": 4,
               },
               "sa": {
                 "a": 0,
                 "k": 0,
-                "ix": 5
+                "ix": 5,
               },
-              "nm": "Transform"
-            }
+              "nm": "Transform",
+            },
           ],
           "nm": "null",
           "np": 2,
           "cix": 2,
           "ix": 1,
           "mn": "ADBE Vector Group",
-          "hd": false
-        }
+          "hd": false,
+        },
       ],
       "ip": 0,
       "op": 848,
       "st": 0,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -275,12 +275,12 @@ export const cloudyDay = {
         "o": {
           "a": 0,
           "k": 0,
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 1,
@@ -288,47 +288,47 @@ export const cloudyDay = {
             {
               "i": {
                 "x": 0.36,
-                "y": 1
+                "y": 1,
               },
               "o": {
                 "x": 0.22,
-                "y": 0.594
+                "y": 0.594,
               },
               "n": "0p36_1_0p22_0p594",
               "t": 1,
               "s": [
                 343,
                 1069,
-                0
+                0,
               ],
               "e": [
                 895,
                 1069,
-                0
+                0,
               ],
               "to": [
                 92,
                 0,
-                0
+                0,
               ],
               "ti": [
                 -92,
                 0,
-                0
-              ]
+                0,
+              ],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             0,
             0,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 1,
@@ -338,53 +338,53 @@ export const cloudyDay = {
                 "x": [
                   0.36,
                   0.36,
-                  0.36
+                  0.36,
                 ],
                 "y": [
                   1,
                   1,
-                  1
-                ]
+                  1,
+                ],
               },
               "o": {
                 "x": [
                   0.22,
                   0.22,
-                  0.22
+                  0.22,
                 ],
                 "y": [
                   0,
                   0,
-                  0
-                ]
+                  0,
+                ],
               },
               "n": [
                 "0p36_1_0p22_0",
                 "0p36_1_0p22_0",
-                "0p36_1_0p22_0"
+                "0p36_1_0p22_0",
               ],
               "t": 1,
               "s": [
                 70,
                 70,
-                100
+                100,
               ],
               "e": [
                 70,
                 70,
-                100
-              ]
+                100,
+              ],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "ip": -33,
       "op": 849,
       "st": -2,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -396,12 +396,12 @@ export const cloudyDay = {
         "o": {
           "a": 0,
           "k": 0,
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 1,
@@ -409,47 +409,47 @@ export const cloudyDay = {
             {
               "i": {
                 "x": 0.36,
-                "y": 1
+                "y": 1,
               },
               "o": {
                 "x": 0.22,
-                "y": 1
+                "y": 1,
               },
               "n": "0p36_1_0p22_1",
               "t": 1,
               "s": [
                 422,
                 750,
-                0
+                0,
               ],
               "e": [
                 750,
                 750,
-                0
+                0,
               ],
               "to": [
                 54.6666679382324,
                 0,
-                0
+                0,
               ],
               "ti": [
                 -54.6666679382324,
                 0,
-                0
-              ]
+                0,
+              ],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             0,
             0,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 1,
@@ -459,53 +459,53 @@ export const cloudyDay = {
                 "x": [
                   0.36,
                   0.36,
-                  0.36
+                  0.36,
                 ],
                 "y": [
                   1,
                   1,
-                  1
-                ]
+                  1,
+                ],
               },
               "o": {
                 "x": [
                   0.22,
                   0.22,
-                  0.22
+                  0.22,
                 ],
                 "y": [
                   0,
                   0,
-                  0
-                ]
+                  0,
+                ],
               },
               "n": [
                 "0p36_1_0p22_0",
                 "0p36_1_0p22_0",
-                "0p36_1_0p22_0"
+                "0p36_1_0p22_0",
               ],
               "t": 1,
               "s": [
                 100,
                 100,
-                100
+                100,
               ],
               "e": [
                 100,
                 100,
-                100
-              ]
+                100,
+              ],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "ip": -33,
       "op": 849,
       "st": -2,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -521,53 +521,53 @@ export const cloudyDay = {
             {
               "i": {
                 "x": [0.518],
-                "y": [0.936]
+                "y": [0.936],
               },
               "o": {
                 "x": [0.294],
-                "y": [0.52]
+                "y": [0.52],
               },
               "n": ["0p518_0p936_0p294_0p52"],
               "t": 1,
               "s": [0],
-              "e": [100]
+              "e": [100],
             },
-            {"t": 37}
+            {"t": 37},
           ],
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 0,
           "k": [
             -250,
             -500,
-            0
+            0,
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             0,
             0,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             100,
             100,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "shapes": [
@@ -581,21 +581,21 @@ export const cloudyDay = {
                 "a": 0,
                 "k": [
                   24,
-                  24
+                  24,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "p": {
                 "a": 0,
                 "k": [
                   12,
-                  12
+                  12,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "nm": "Ellipse Path 1",
               "mn": "ADBE Vector Shape - Ellipse",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "fl",
@@ -605,26 +605,26 @@ export const cloudyDay = {
                   0.945097982883,
                   0.952941000462,
                   0.956862986088,
-                  1
+                  1,
                 ],
-                "ix": 4
+                "ix": 4,
               },
               "o": {
                 "a": 0,
                 "k": 80,
-                "ix": 5
+                "ix": 5,
               },
               "r": 2,
               "nm": "Fill 1",
               "mn": "ADBE Vector Graphic - Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "gf",
               "o": {
                 "a": 0,
                 "k": 50,
-                "ix": 10
+                "ix": 10,
               },
               "r": 2,
               "g": {
@@ -649,31 +649,31 @@ export const cloudyDay = {
                     0.5,
                     0.1,
                     1,
-                    0.2
+                    0.2,
                   ],
-                  "ix": 9
-                }
+                  "ix": 9,
+                },
               },
               "s": {
                 "a": 0,
                 "k": [
                   11.385,
-                  13.566
+                  13.566,
                 ],
-                "ix": 5
+                "ix": 5,
               },
               "e": {
                 "a": 0,
                 "k": [
                   23.401,
-                  24.13
+                  24.13,
                 ],
-                "ix": 6
+                "ix": 6,
               },
               "t": 1,
               "nm": "Gradient Fill 1",
               "mn": "ADBE Vector Graphic - G-Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "tr",
@@ -681,61 +681,61 @@ export const cloudyDay = {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "a": {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 1
+                "ix": 1,
               },
               "s": {
                 "a": 0,
                 "k": [
                   3125,
-                  3125
+                  3125,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "r": {
                 "a": 0,
                 "k": 0,
-                "ix": 6
+                "ix": 6,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 7
+                "ix": 7,
               },
               "sk": {
                 "a": 0,
                 "k": 0,
-                "ix": 4
+                "ix": 4,
               },
               "sa": {
                 "a": 0,
                 "k": 0,
-                "ix": 5
+                "ix": 5,
               },
-              "nm": "Transform"
-            }
+              "nm": "Transform",
+            },
           ],
           "nm": "Oval",
           "np": 3,
           "cix": 2,
           "ix": 1,
           "mn": "ADBE Vector Group",
-          "hd": false
-        }
+          "hd": false,
+        },
       ],
       "ip": -58,
       "op": 842,
       "st": -6,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -750,25 +750,25 @@ export const cloudyDay = {
             {
               "i": {
                 "x": [0.36],
-                "y": [1]
+                "y": [1],
               },
               "o": {
                 "x": [0.22],
-                "y": [99]
+                "y": [99],
               },
               "n": ["0p36_1_0p22_99"],
               "t": 1,
               "s": [100],
-              "e": [100]
+              "e": [100],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 1,
@@ -776,57 +776,57 @@ export const cloudyDay = {
             {
               "i": {
                 "x": 0.36,
-                "y": 1
+                "y": 1,
               },
               "o": {
                 "x": 0.22,
-                "y": 1
+                "y": 1,
               },
               "n": "0p36_1_0p22_1",
               "t": 1,
               "s": [
                 487.75,
                 237,
-                0
+                0,
               ],
               "e": [
                 232.75,
                 237,
-                0
+                0,
               ],
               "to": [
                 -42.5,
                 0,
-                0
+                0,
               ],
               "ti": [
                 42.5,
                 0,
-                0
-              ]
+                0,
+              ],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             -496,
             -494,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             100,
             100,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "shapes": [
@@ -840,21 +840,21 @@ export const cloudyDay = {
                 "a": 0,
                 "k": [
                   365.547,
-                  365.547
+                  365.547,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "p": {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "nm": "Ellipse Path 1",
               "mn": "ADBE Vector Shape - Ellipse",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "tr",
@@ -862,55 +862,55 @@ export const cloudyDay = {
                 "a": 0,
                 "k": [
                   -499.359,
-                  -499.43
+                  -499.43,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "a": {
                 "a": 0,
                 "k": [
                   -209.552,
-                  -208.357
+                  -208.357,
                 ],
-                "ix": 1
+                "ix": 1,
               },
               "s": {
                 "a": 0,
                 "k": [
                   131.101,
-                  128.967
+                  128.967,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "r": {
                 "a": 0,
                 "k": 0,
-                "ix": 6
+                "ix": 6,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 7
+                "ix": 7,
               },
               "sk": {
                 "a": 0,
                 "k": 0,
-                "ix": 4
+                "ix": 4,
               },
               "sa": {
                 "a": 0,
                 "k": 0,
-                "ix": 5
+                "ix": 5,
               },
-              "nm": "Transform"
-            }
+              "nm": "Transform",
+            },
           ],
           "nm": "Ellipse 1",
           "np": 1,
           "cix": 2,
           "ix": 1,
           "mn": "ADBE Vector Group",
-          "hd": false
+          "hd": false,
         },
         {
           "ty": "st",
@@ -920,14 +920,14 @@ export const cloudyDay = {
               0.98431372549,
               0.737254901961,
               0.01568627451,
-              1
+              1,
             ],
-            "ix": 3
+            "ix": 3,
           },
           "o": {
             "a": 0,
             "k": 100,
-            "ix": 4
+            "ix": 4,
           },
           "w": {
             "a": 1,
@@ -935,33 +935,33 @@ export const cloudyDay = {
               {
                 "i": {
                   "x": [0.36],
-                  "y": [1]
+                  "y": [1],
                 },
                 "o": {
                   "x": [0.22],
-                  "y": [1]
+                  "y": [1],
                 },
                 "n": ["0p36_1_0p22_1"],
                 "t": 1,
                 "s": [1],
-                "e": [30]
+                "e": [30],
               },
-              {"t": 75}
+              {"t": 75},
             ],
-            "ix": 5
+            "ix": 5,
           },
           "lc": 1,
           "lj": 1,
           "ml": 4,
           "nm": "Stroke 1",
           "mn": "ADBE Vector Graphic - Stroke",
-          "hd": false
-        }
+          "hd": false,
+        },
       ],
       "ip": 1,
       "op": 901,
       "st": -1,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -975,40 +975,40 @@ export const cloudyDay = {
         "o": {
           "a": 0,
           "k": 100,
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 0,
           "k": [
             -250,
             -500,
-            0
+            0,
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             0,
             0,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             100,
             100,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "shapes": [
@@ -1022,21 +1022,21 @@ export const cloudyDay = {
                 "a": 0,
                 "k": [
                   24,
-                  24
+                  24,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "p": {
                 "a": 0,
                 "k": [
                   12,
-                  12
+                  12,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "nm": "Ellipse Path 1",
               "mn": "ADBE Vector Shape - Ellipse",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "fl",
@@ -1046,26 +1046,26 @@ export const cloudyDay = {
                   0.945097982883,
                   0.952941000462,
                   0.956862986088,
-                  1
+                  1,
                 ],
-                "ix": 4
+                "ix": 4,
               },
               "o": {
                 "a": 0,
                 "k": 80,
-                "ix": 5
+                "ix": 5,
               },
               "r": 2,
               "nm": "Fill 1",
               "mn": "ADBE Vector Graphic - Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "gf",
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 10
+                "ix": 10,
               },
               "r": 2,
               "g": {
@@ -1090,31 +1090,31 @@ export const cloudyDay = {
                     0.5,
                     0.1,
                     1,
-                    0.2
+                    0.2,
                   ],
-                  "ix": 9
-                }
+                  "ix": 9,
+                },
               },
               "s": {
                 "a": 0,
                 "k": [
                   11.385,
-                  13.566
+                  13.566,
                 ],
-                "ix": 5
+                "ix": 5,
               },
               "e": {
                 "a": 0,
                 "k": [
                   23.401,
-                  24.13
+                  24.13,
                 ],
-                "ix": 6
+                "ix": 6,
               },
               "t": 1,
               "nm": "Gradient Fill 1",
               "mn": "ADBE Vector Graphic - G-Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "tr",
@@ -1122,61 +1122,61 @@ export const cloudyDay = {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "a": {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 1
+                "ix": 1,
               },
               "s": {
                 "a": 0,
                 "k": [
                   3125,
-                  3125
+                  3125,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "r": {
                 "a": 0,
                 "k": 0,
-                "ix": 6
+                "ix": 6,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 7
+                "ix": 7,
               },
               "sk": {
                 "a": 0,
                 "k": 0,
-                "ix": 4
+                "ix": 4,
               },
               "sa": {
                 "a": 0,
                 "k": 0,
-                "ix": 5
+                "ix": 5,
               },
-              "nm": "Transform"
-            }
+              "nm": "Transform",
+            },
           ],
           "nm": "Oval",
           "np": 3,
           "cix": 2,
           "ix": 1,
           "mn": "ADBE Vector Group",
-          "hd": false
-        }
+          "hd": false,
+        },
       ],
       "ip": -58,
       "op": 842,
       "st": -6,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -1193,53 +1193,53 @@ export const cloudyDay = {
             {
               "i": {
                 "x": [0.36],
-                "y": [1]
+                "y": [1],
               },
               "o": {
                 "x": [0.22],
-                "y": [1]
+                "y": [1],
               },
               "n": ["0p36_1_0p22_1"],
               "t": 1,
               "s": [0],
-              "e": [100]
+              "e": [100],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 0,
           "k": [
             -84,
             -62.5,
-            0
+            0,
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             416,
             0,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             100,
             100,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "shapes": [
@@ -1256,52 +1256,52 @@ export const cloudyDay = {
                   "i": [
                     [
                       5.52,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      -5.52
+                      -5.52,
                     ],
                     [
                       0.01,
-                      -0.02
-                    ]
+                      -0.02,
+                    ],
                   ],
                   "o": [
                     [
                       -5.52,
-                      0
+                      0,
                     ],
                     [
                       0.01,
-                      -0.02
+                      -0.02,
                     ],
                     [
                       0,
-                      -5.52
-                    ]
+                      -5.52,
+                    ],
                   ],
                   "v": [
                     [
                       10,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      10
+                      10,
                     ],
                     [
                       20,
-                      10
-                    ]
+                      10,
+                    ],
                   ],
-                  "c": true
+                  "c": true,
                 },
-                "ix": 2
+                "ix": 2,
               },
               "nm": "Path 1",
               "mn": "ADBE Vector Shape - Group",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "fl",
@@ -1311,19 +1311,19 @@ export const cloudyDay = {
                   0.945097982883,
                   0.952941000462,
                   0.956862986088,
-                  1
+                  1,
                 ],
-                "ix": 4
+                "ix": 4,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 5
+                "ix": 5,
               },
               "r": 1,
               "nm": "Fill 1",
               "mn": "ADBE Vector Graphic - Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "tr",
@@ -1331,62 +1331,62 @@ export const cloudyDay = {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "a": {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 1
+                "ix": 1,
               },
               "s": {
                 "a": 0,
                 "k": [
                   3125,
-                  3125
+                  3125,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "r": {
                 "a": 0,
                 "k": 0,
-                "ix": 6
+                "ix": 6,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 7
+                "ix": 7,
               },
               "sk": {
                 "a": 0,
                 "k": 0,
-                "ix": 4
+                "ix": 4,
               },
               "sa": {
                 "a": 0,
                 "k": 0,
-                "ix": 5
+                "ix": 5,
               },
-              "nm": "Transform"
-            }
+              "nm": "Transform",
+            },
           ],
           "nm": "Group 1",
           "np": 2,
           "cix": 2,
           "ix": 1,
           "mn": "ADBE Vector Group",
-          "hd": false
-        }
+          "hd": false,
+        },
       ],
       "ip": -52,
       "op": 848,
       "st": -70,
-      "bm": 0
-    }
+      "bm": 0,
+    },
   ],
-  "markers": []
+  "markers": [],
 };
