@@ -19,40 +19,40 @@ export const isolatedStormsDay = {
         "o": {
           "a": 0,
           "k": 100,
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 0,
           "k": [
             938.5,
             1000,
-            0
+            0,
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             88,
             117,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             213,
             213,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "shapes": [
@@ -69,100 +69,100 @@ export const isolatedStormsDay = {
                   "i": [
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
-                    ]
+                      0,
+                    ],
                   ],
                   "o": [
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
-                    ]
+                      0,
+                    ],
                   ],
                   "v": [
                     [
                       -29.333,
-                      -117
+                      -117,
                     ],
                     [
                       -88,
-                      0
+                      0,
                     ],
                     [
                       -35.2,
-                      0
+                      0,
                     ],
                     [
                       -41.741,
-                      117
+                      117,
                     ],
                     [
                       87.325,
-                      -29.25
+                      -29.25,
                     ],
                     [
                       41.052,
-                      -29.25
+                      -29.25,
                     ],
                     [
                       88,
-                      -117
-                    ]
+                      -117,
+                    ],
                   ],
-                  "c": true
+                  "c": true,
                 },
-                "ix": 2
+                "ix": 2,
               },
               "nm": "Path 2",
               "mn": "ADBE Vector Shape - Group",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "fl",
@@ -172,9 +172,9 @@ export const isolatedStormsDay = {
                   1,
                   1,
                   1,
-                  1
+                  1,
                 ],
-                "ix": 4
+                "ix": 4,
               },
               "o": {
                 "a": 1,
@@ -182,39 +182,39 @@ export const isolatedStormsDay = {
                   {
                     "i": {
                       "x": [0.833],
-                      "y": [0.833]
+                      "y": [0.833],
                     },
                     "o": {
                       "x": [0.167],
-                      "y": [0.167]
+                      "y": [0.167],
                     },
                     "n": ["0p833_0p833_0p167_0p167"],
                     "t": 55,
                     "s": [1],
-                    "e": [100]
+                    "e": [100],
                   },
                   {
                     "i": {
                       "x": [0.833],
-                      "y": [0.833]
+                      "y": [0.833],
                     },
                     "o": {
                       "x": [0.167],
-                      "y": [0.167]
+                      "y": [0.167],
                     },
                     "n": ["0p833_0p833_0p167_0p167"],
                     "t": 59,
                     "s": [100],
-                    "e": [3]
+                    "e": [3],
                   },
-                  {"t": 104}
+                  {"t": 104},
                 ],
-                "ix": 5
+                "ix": 5,
               },
               "r": 1,
               "nm": "Fill 1",
               "mn": "ADBE Vector Graphic - Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "tr",
@@ -222,61 +222,61 @@ export const isolatedStormsDay = {
                 "a": 0,
                 "k": [
                   88,
-                  117
+                  117,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "a": {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 1
+                "ix": 1,
               },
               "s": {
                 "a": 0,
                 "k": [
                   100,
-                  100
+                  100,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "r": {
                 "a": 0,
                 "k": 0,
-                "ix": 6
+                "ix": 6,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 7
+                "ix": 7,
               },
               "sk": {
                 "a": 0,
                 "k": 0,
-                "ix": 4
+                "ix": 4,
               },
               "sa": {
                 "a": 0,
                 "k": 0,
-                "ix": 5
+                "ix": 5,
               },
-              "nm": "Transform"
-            }
+              "nm": "Transform",
+            },
           ],
           "nm": "Group 1",
           "np": 2,
           "cix": 2,
           "ix": 1,
           "mn": "ADBE Vector Group",
-          "hd": false
-        }
+          "hd": false,
+        },
       ],
       "ip": 0,
       "op": 848,
       "st": 0,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -291,53 +291,53 @@ export const isolatedStormsDay = {
             {
               "i": {
                 "x": [0.833],
-                "y": [0.833]
+                "y": [0.833],
               },
               "o": {
                 "x": [0.167],
-                "y": [0.167]
+                "y": [0.167],
               },
               "n": ["0p833_0p833_0p167_0p167"],
               "t": 59,
               "s": [0],
-              "e": [100]
+              "e": [100],
             },
-            {"t": 66}
+            {"t": 66},
           ],
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 0,
           "k": [
             938.5,
             1000,
-            0
+            0,
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             88,
             117,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             213,
             213,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "shapes": [
@@ -354,100 +354,100 @@ export const isolatedStormsDay = {
                   "i": [
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
-                    ]
+                      0,
+                    ],
                   ],
                   "o": [
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
-                    ]
+                      0,
+                    ],
                   ],
                   "v": [
                     [
                       -29.333,
-                      -117
+                      -117,
                     ],
                     [
                       -88,
-                      0
+                      0,
                     ],
                     [
                       -35.2,
-                      0
+                      0,
                     ],
                     [
                       -41.741,
-                      117
+                      117,
                     ],
                     [
                       87.325,
-                      -29.25
+                      -29.25,
                     ],
                     [
                       41.052,
-                      -29.25
+                      -29.25,
                     ],
                     [
                       88,
-                      -117
-                    ]
+                      -117,
+                    ],
                   ],
-                  "c": true
+                  "c": true,
                 },
-                "ix": 2
+                "ix": 2,
               },
               "nm": "Path 2",
               "mn": "ADBE Vector Shape - Group",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "fl",
@@ -457,19 +457,19 @@ export const isolatedStormsDay = {
                   0.98431372549,
                   0.737254901961,
                   0.01568627451,
-                  1
+                  1,
                 ],
-                "ix": 4
+                "ix": 4,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 5
+                "ix": 5,
               },
               "r": 1,
               "nm": "Fill 1",
               "mn": "ADBE Vector Graphic - Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "tr",
@@ -477,61 +477,61 @@ export const isolatedStormsDay = {
                 "a": 0,
                 "k": [
                   88,
-                  117
+                  117,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "a": {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 1
+                "ix": 1,
               },
               "s": {
                 "a": 0,
                 "k": [
                   100,
-                  100
+                  100,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "r": {
                 "a": 0,
                 "k": 0,
-                "ix": 6
+                "ix": 6,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 7
+                "ix": 7,
               },
               "sk": {
                 "a": 0,
                 "k": 0,
-                "ix": 4
+                "ix": 4,
               },
               "sa": {
                 "a": 0,
                 "k": 0,
-                "ix": 5
+                "ix": 5,
               },
-              "nm": "Transform"
-            }
+              "nm": "Transform",
+            },
           ],
           "nm": "Group 1",
           "np": 2,
           "cix": 2,
           "ix": 1,
           "mn": "ADBE Vector Group",
-          "hd": false
-        }
+          "hd": false,
+        },
       ],
       "ip": 0,
       "op": 848,
       "st": 0,
-      "bm": 1
+      "bm": 1,
     },
     {
       "ddd": 0,
@@ -543,12 +543,12 @@ export const isolatedStormsDay = {
         "o": {
           "a": 0,
           "k": 0,
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 1,
@@ -556,47 +556,47 @@ export const isolatedStormsDay = {
             {
               "i": {
                 "x": 0.36,
-                "y": 1
+                "y": 1,
               },
               "o": {
                 "x": 0.22,
-                "y": 1
+                "y": 1,
               },
               "n": "0p36_1_0p22_1",
               "t": 1,
               "s": [
                 422,
                 750,
-                0
+                0,
               ],
               "e": [
                 750,
                 750,
-                0
+                0,
               ],
               "to": [
                 54.6666679382324,
                 0,
-                0
+                0,
               ],
               "ti": [
                 -54.6666679382324,
                 0,
-                0
-              ]
+                0,
+              ],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             0,
             0,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 1,
@@ -606,53 +606,53 @@ export const isolatedStormsDay = {
                 "x": [
                   0.36,
                   0.36,
-                  0.36
+                  0.36,
                 ],
                 "y": [
                   1,
                   1,
-                  1
-                ]
+                  1,
+                ],
               },
               "o": {
                 "x": [
                   0.22,
                   0.22,
-                  0.22
+                  0.22,
                 ],
                 "y": [
                   0,
                   0,
-                  0
-                ]
+                  0,
+                ],
               },
               "n": [
                 "0p36_1_0p22_0",
                 "0p36_1_0p22_0",
-                "0p36_1_0p22_0"
+                "0p36_1_0p22_0",
               ],
               "t": 1,
               "s": [
                 100,
                 100,
-                100
+                100,
               ],
               "e": [
                 100,
                 100,
-                100
-              ]
+                100,
+              ],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "ip": -33,
       "op": 849,
       "st": -2,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -668,53 +668,53 @@ export const isolatedStormsDay = {
             {
               "i": {
                 "x": [0.518],
-                "y": [0.936]
+                "y": [0.936],
               },
               "o": {
                 "x": [0.294],
-                "y": [0.52]
+                "y": [0.52],
               },
               "n": ["0p518_0p936_0p294_0p52"],
               "t": 1,
               "s": [0],
-              "e": [100]
+              "e": [100],
             },
-            {"t": 37}
+            {"t": 37},
           ],
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 0,
           "k": [
             -250,
             -500,
-            0
+            0,
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             0,
             0,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             100,
             100,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "shapes": [
@@ -728,21 +728,21 @@ export const isolatedStormsDay = {
                 "a": 0,
                 "k": [
                   24,
-                  24
+                  24,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "p": {
                 "a": 0,
                 "k": [
                   12,
-                  12
+                  12,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "nm": "Ellipse Path 1",
               "mn": "ADBE Vector Shape - Ellipse",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "fl",
@@ -752,26 +752,26 @@ export const isolatedStormsDay = {
                   0.945097982883,
                   0.952941000462,
                   0.956862986088,
-                  1
+                  1,
                 ],
-                "ix": 4
+                "ix": 4,
               },
               "o": {
                 "a": 0,
                 "k": 80,
-                "ix": 5
+                "ix": 5,
               },
               "r": 2,
               "nm": "Fill 1",
               "mn": "ADBE Vector Graphic - Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "gf",
               "o": {
                 "a": 0,
                 "k": 50,
-                "ix": 10
+                "ix": 10,
               },
               "r": 2,
               "g": {
@@ -796,31 +796,31 @@ export const isolatedStormsDay = {
                     0.5,
                     0.1,
                     1,
-                    0.2
+                    0.2,
                   ],
-                  "ix": 9
-                }
+                  "ix": 9,
+                },
               },
               "s": {
                 "a": 0,
                 "k": [
                   11.385,
-                  13.566
+                  13.566,
                 ],
-                "ix": 5
+                "ix": 5,
               },
               "e": {
                 "a": 0,
                 "k": [
                   23.401,
-                  24.13
+                  24.13,
                 ],
-                "ix": 6
+                "ix": 6,
               },
               "t": 1,
               "nm": "Gradient Fill 1",
               "mn": "ADBE Vector Graphic - G-Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "tr",
@@ -828,61 +828,61 @@ export const isolatedStormsDay = {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "a": {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 1
+                "ix": 1,
               },
               "s": {
                 "a": 0,
                 "k": [
                   3125,
-                  3125
+                  3125,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "r": {
                 "a": 0,
                 "k": 0,
-                "ix": 6
+                "ix": 6,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 7
+                "ix": 7,
               },
               "sk": {
                 "a": 0,
                 "k": 0,
-                "ix": 4
+                "ix": 4,
               },
               "sa": {
                 "a": 0,
                 "k": 0,
-                "ix": 5
+                "ix": 5,
               },
-              "nm": "Transform"
-            }
+              "nm": "Transform",
+            },
           ],
           "nm": "Oval",
           "np": 3,
           "cix": 2,
           "ix": 1,
           "mn": "ADBE Vector Group",
-          "hd": false
-        }
+          "hd": false,
+        },
       ],
       "ip": -58,
       "op": 842,
       "st": -6,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -897,25 +897,25 @@ export const isolatedStormsDay = {
             {
               "i": {
                 "x": [0.36],
-                "y": [1]
+                "y": [1],
               },
               "o": {
                 "x": [0.22],
-                "y": [99]
+                "y": [99],
               },
               "n": ["0p36_1_0p22_99"],
               "t": 1,
               "s": [100],
-              "e": [100]
+              "e": [100],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 1,
@@ -923,57 +923,57 @@ export const isolatedStormsDay = {
             {
               "i": {
                 "x": 0.36,
-                "y": 1
+                "y": 1,
               },
               "o": {
                 "x": 0.22,
-                "y": 1
+                "y": 1,
               },
               "n": "0p36_1_0p22_1",
               "t": 1,
               "s": [
                 487.75,
                 237,
-                0
+                0,
               ],
               "e": [
                 232.75,
                 237,
-                0
+                0,
               ],
               "to": [
                 -42.5,
                 0,
-                0
+                0,
               ],
               "ti": [
                 42.5,
                 0,
-                0
-              ]
+                0,
+              ],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             -496,
             -494,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             100,
             100,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "shapes": [
@@ -987,21 +987,21 @@ export const isolatedStormsDay = {
                 "a": 0,
                 "k": [
                   365.547,
-                  365.547
+                  365.547,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "p": {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "nm": "Ellipse Path 1",
               "mn": "ADBE Vector Shape - Ellipse",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "tr",
@@ -1009,55 +1009,55 @@ export const isolatedStormsDay = {
                 "a": 0,
                 "k": [
                   -499.359,
-                  -499.43
+                  -499.43,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "a": {
                 "a": 0,
                 "k": [
                   -209.552,
-                  -208.357
+                  -208.357,
                 ],
-                "ix": 1
+                "ix": 1,
               },
               "s": {
                 "a": 0,
                 "k": [
                   131.101,
-                  128.967
+                  128.967,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "r": {
                 "a": 0,
                 "k": 0,
-                "ix": 6
+                "ix": 6,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 7
+                "ix": 7,
               },
               "sk": {
                 "a": 0,
                 "k": 0,
-                "ix": 4
+                "ix": 4,
               },
               "sa": {
                 "a": 0,
                 "k": 0,
-                "ix": 5
+                "ix": 5,
               },
-              "nm": "Transform"
-            }
+              "nm": "Transform",
+            },
           ],
           "nm": "Ellipse 1",
           "np": 1,
           "cix": 2,
           "ix": 1,
           "mn": "ADBE Vector Group",
-          "hd": false
+          "hd": false,
         },
         {
           "ty": "st",
@@ -1067,14 +1067,14 @@ export const isolatedStormsDay = {
               0.98431372549,
               0.737254901961,
               0.01568627451,
-              1
+              1,
             ],
-            "ix": 3
+            "ix": 3,
           },
           "o": {
             "a": 0,
             "k": 100,
-            "ix": 4
+            "ix": 4,
           },
           "w": {
             "a": 1,
@@ -1082,33 +1082,33 @@ export const isolatedStormsDay = {
               {
                 "i": {
                   "x": [0.36],
-                  "y": [1]
+                  "y": [1],
                 },
                 "o": {
                   "x": [0.22],
-                  "y": [1]
+                  "y": [1],
                 },
                 "n": ["0p36_1_0p22_1"],
                 "t": 1,
                 "s": [1],
-                "e": [30]
+                "e": [30],
               },
-              {"t": 75}
+              {"t": 75},
             ],
-            "ix": 5
+            "ix": 5,
           },
           "lc": 1,
           "lj": 1,
           "ml": 4,
           "nm": "Stroke 1",
           "mn": "ADBE Vector Graphic - Stroke",
-          "hd": false
-        }
+          "hd": false,
+        },
       ],
       "ip": 1,
       "op": 901,
       "st": -1,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -1122,40 +1122,40 @@ export const isolatedStormsDay = {
         "o": {
           "a": 0,
           "k": 100,
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 0,
           "k": [
             -250,
             -500,
-            0
+            0,
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             0,
             0,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             100,
             100,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "shapes": [
@@ -1169,21 +1169,21 @@ export const isolatedStormsDay = {
                 "a": 0,
                 "k": [
                   24,
-                  24
+                  24,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "p": {
                 "a": 0,
                 "k": [
                   12,
-                  12
+                  12,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "nm": "Ellipse Path 1",
               "mn": "ADBE Vector Shape - Ellipse",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "fl",
@@ -1193,26 +1193,26 @@ export const isolatedStormsDay = {
                   0.945097982883,
                   0.952941000462,
                   0.956862986088,
-                  1
+                  1,
                 ],
-                "ix": 4
+                "ix": 4,
               },
               "o": {
                 "a": 0,
                 "k": 80,
-                "ix": 5
+                "ix": 5,
               },
               "r": 2,
               "nm": "Fill 1",
               "mn": "ADBE Vector Graphic - Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "gf",
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 10
+                "ix": 10,
               },
               "r": 2,
               "g": {
@@ -1237,31 +1237,31 @@ export const isolatedStormsDay = {
                     0.5,
                     0.1,
                     1,
-                    0.2
+                    0.2,
                   ],
-                  "ix": 9
-                }
+                  "ix": 9,
+                },
               },
               "s": {
                 "a": 0,
                 "k": [
                   11.385,
-                  13.566
+                  13.566,
                 ],
-                "ix": 5
+                "ix": 5,
               },
               "e": {
                 "a": 0,
                 "k": [
                   23.401,
-                  24.13
+                  24.13,
                 ],
-                "ix": 6
+                "ix": 6,
               },
               "t": 1,
               "nm": "Gradient Fill 1",
               "mn": "ADBE Vector Graphic - G-Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "tr",
@@ -1269,61 +1269,61 @@ export const isolatedStormsDay = {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "a": {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 1
+                "ix": 1,
               },
               "s": {
                 "a": 0,
                 "k": [
                   3125,
-                  3125
+                  3125,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "r": {
                 "a": 0,
                 "k": 0,
-                "ix": 6
+                "ix": 6,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 7
+                "ix": 7,
               },
               "sk": {
                 "a": 0,
                 "k": 0,
-                "ix": 4
+                "ix": 4,
               },
               "sa": {
                 "a": 0,
                 "k": 0,
-                "ix": 5
+                "ix": 5,
               },
-              "nm": "Transform"
-            }
+              "nm": "Transform",
+            },
           ],
           "nm": "Oval",
           "np": 3,
           "cix": 2,
           "ix": 1,
           "mn": "ADBE Vector Group",
-          "hd": false
-        }
+          "hd": false,
+        },
       ],
       "ip": -58,
       "op": 842,
       "st": -6,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -1340,53 +1340,53 @@ export const isolatedStormsDay = {
             {
               "i": {
                 "x": [0.36],
-                "y": [1]
+                "y": [1],
               },
               "o": {
                 "x": [0.22],
-                "y": [1]
+                "y": [1],
               },
               "n": ["0p36_1_0p22_1"],
               "t": 1,
               "s": [0],
-              "e": [100]
+              "e": [100],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 0,
           "k": [
             -84,
             -62.5,
-            0
+            0,
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             416,
             0,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             100,
             100,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "shapes": [
@@ -1403,52 +1403,52 @@ export const isolatedStormsDay = {
                   "i": [
                     [
                       5.52,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      -5.52
+                      -5.52,
                     ],
                     [
                       0.01,
-                      -0.02
-                    ]
+                      -0.02,
+                    ],
                   ],
                   "o": [
                     [
                       -5.52,
-                      0
+                      0,
                     ],
                     [
                       0.01,
-                      -0.02
+                      -0.02,
                     ],
                     [
                       0,
-                      -5.52
-                    ]
+                      -5.52,
+                    ],
                   ],
                   "v": [
                     [
                       10,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      10
+                      10,
                     ],
                     [
                       20,
-                      10
-                    ]
+                      10,
+                    ],
                   ],
-                  "c": true
+                  "c": true,
                 },
-                "ix": 2
+                "ix": 2,
               },
               "nm": "Path 1",
               "mn": "ADBE Vector Shape - Group",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "fl",
@@ -1458,19 +1458,19 @@ export const isolatedStormsDay = {
                   0.945097982883,
                   0.952941000462,
                   0.956862986088,
-                  1
+                  1,
                 ],
-                "ix": 4
+                "ix": 4,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 5
+                "ix": 5,
               },
               "r": 1,
               "nm": "Fill 1",
               "mn": "ADBE Vector Graphic - Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "tr",
@@ -1478,62 +1478,62 @@ export const isolatedStormsDay = {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "a": {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 1
+                "ix": 1,
               },
               "s": {
                 "a": 0,
                 "k": [
                   3125,
-                  3125
+                  3125,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "r": {
                 "a": 0,
                 "k": 0,
-                "ix": 6
+                "ix": 6,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 7
+                "ix": 7,
               },
               "sk": {
                 "a": 0,
                 "k": 0,
-                "ix": 4
+                "ix": 4,
               },
               "sa": {
                 "a": 0,
                 "k": 0,
-                "ix": 5
+                "ix": 5,
               },
-              "nm": "Transform"
-            }
+              "nm": "Transform",
+            },
           ],
           "nm": "Group 1",
           "np": 2,
           "cix": 2,
           "ix": 1,
           "mn": "ADBE Vector Group",
-          "hd": false
-        }
+          "hd": false,
+        },
       ],
       "ip": -52,
       "op": 848,
       "st": -70,
-      "bm": 0
-    }
+      "bm": 0,
+    },
   ],
-  "markers": []
+  "markers": [],
 };
