@@ -21,40 +21,40 @@ export const isolatedStormsShowersDay = {
             "o": {
               "a": 0,
               "k": 100,
-              "ix": 11
+              "ix": 11,
             },
             "r": {
               "a": 0,
               "k": -45,
-              "ix": 10
+              "ix": 10,
             },
             "p": {
               "a": 0,
               "k": [
                 -15,
                 122.5,
-                0
+                0,
               ],
-              "ix": 2
+              "ix": 2,
             },
             "a": {
               "a": 0,
               "k": [
                 0,
                 0,
-                0
+                0,
               ],
-              "ix": 1
+              "ix": 1,
             },
             "s": {
               "a": 0,
               "k": [
                 1062.5,
                 1062.5,
-                100
+                100,
               ],
-              "ix": 6
-            }
+              "ix": 6,
+            },
           },
           "ao": 0,
           "shapes": [
@@ -71,21 +71,21 @@ export const isolatedStormsShowersDay = {
                         "a": 0,
                         "k": [
                           16,
-                          16
+                          16,
                         ],
-                        "ix": 2
+                        "ix": 2,
                       },
                       "p": {
                         "a": 0,
                         "k": [
                           8,
-                          8
+                          8,
                         ],
-                        "ix": 3
+                        "ix": 3,
                       },
                       "nm": "Ellipse Path 1",
                       "mn": "ADBE Vector Shape - Ellipse",
-                      "hd": false
+                      "hd": false,
                     },
                     {
                       "ty": "tr",
@@ -93,55 +93,55 @@ export const isolatedStormsShowersDay = {
                         "a": 0,
                         "k": [
                           8,
-                          8
+                          8,
                         ],
-                        "ix": 2
+                        "ix": 2,
                       },
                       "a": {
                         "a": 0,
                         "k": [
                           8,
-                          8
+                          8,
                         ],
-                        "ix": 1
+                        "ix": 1,
                       },
                       "s": {
                         "a": 0,
                         "k": [
                           100,
-                          100
+                          100,
                         ],
-                        "ix": 3
+                        "ix": 3,
                       },
                       "r": {
                         "a": 0,
                         "k": 0,
-                        "ix": 6
+                        "ix": 6,
                       },
                       "o": {
                         "a": 0,
                         "k": 100,
-                        "ix": 7
+                        "ix": 7,
                       },
                       "sk": {
                         "a": 0,
                         "k": 0,
-                        "ix": 4
+                        "ix": 4,
                       },
                       "sa": {
                         "a": 0,
                         "k": 0,
-                        "ix": 5
+                        "ix": 5,
                       },
-                      "nm": "Transform"
-                    }
+                      "nm": "Transform",
+                    },
                   ],
                   "nm": "Oval",
                   "np": 1,
                   "cix": 2,
                   "ix": 1,
                   "mn": "ADBE Vector Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ty": "gr",
@@ -153,26 +153,26 @@ export const isolatedStormsShowersDay = {
                         "a": 0,
                         "k": [
                           8,
-                          8
+                          8,
                         ],
-                        "ix": 2
+                        "ix": 2,
                       },
                       "p": {
                         "a": 0,
                         "k": [
                           12,
-                          4
+                          4,
                         ],
-                        "ix": 3
+                        "ix": 3,
                       },
                       "r": {
                         "a": 0,
                         "k": 0,
-                        "ix": 4
+                        "ix": 4,
                       },
                       "nm": "Rectangle Path 1",
                       "mn": "ADBE Vector Shape - Rect",
-                      "hd": false
+                      "hd": false,
                     },
                     {
                       "ty": "tr",
@@ -180,55 +180,55 @@ export const isolatedStormsShowersDay = {
                         "a": 0,
                         "k": [
                           12,
-                          4
+                          4,
                         ],
-                        "ix": 2
+                        "ix": 2,
                       },
                       "a": {
                         "a": 0,
                         "k": [
                           12,
-                          4
+                          4,
                         ],
-                        "ix": 1
+                        "ix": 1,
                       },
                       "s": {
                         "a": 0,
                         "k": [
                           100,
-                          100
+                          100,
                         ],
-                        "ix": 3
+                        "ix": 3,
                       },
                       "r": {
                         "a": 0,
                         "k": 0,
-                        "ix": 6
+                        "ix": 6,
                       },
                       "o": {
                         "a": 0,
                         "k": 100,
-                        "ix": 7
+                        "ix": 7,
                       },
                       "sk": {
                         "a": 0,
                         "k": 0,
-                        "ix": 4
+                        "ix": 4,
                       },
                       "sa": {
                         "a": 0,
                         "k": 0,
-                        "ix": 5
+                        "ix": 5,
                       },
-                      "nm": "Transform"
-                    }
+                      "nm": "Transform",
+                    },
                   ],
                   "nm": "Rectangle 2",
                   "np": 1,
                   "cix": 2,
                   "ix": 2,
                   "mn": "ADBE Vector Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ty": "fl",
@@ -238,19 +238,19 @@ export const isolatedStormsShowersDay = {
                       0.258823990822,
                       0.521569013596,
                       0.956862986088,
-                      1
+                      1,
                     ],
-                    "ix": 4
+                    "ix": 4,
                   },
                   "o": {
                     "a": 0,
                     "k": 100,
-                    "ix": 5
+                    "ix": 5,
                   },
                   "r": 1,
                   "nm": "Fill 1",
                   "mn": "ADBE Vector Graphic - Fill",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ty": "tr",
@@ -258,64 +258,64 @@ export const isolatedStormsShowersDay = {
                     "a": 0,
                     "k": [
                       0,
-                      0
+                      0,
                     ],
-                    "ix": 2
+                    "ix": 2,
                   },
                   "a": {
                     "a": 0,
                     "k": [
                       0,
-                      0
+                      0,
                     ],
-                    "ix": 1
+                    "ix": 1,
                   },
                   "s": {
                     "a": 0,
                     "k": [
                       100,
-                      100
+                      100,
                     ],
-                    "ix": 3
+                    "ix": 3,
                   },
                   "r": {
                     "a": 0,
                     "k": 0,
-                    "ix": 6
+                    "ix": 6,
                   },
                   "o": {
                     "a": 0,
                     "k": 100,
-                    "ix": 7
+                    "ix": 7,
                   },
                   "sk": {
                     "a": 0,
                     "k": 0,
-                    "ix": 4
+                    "ix": 4,
                   },
                   "sa": {
                     "a": 0,
                     "k": 0,
-                    "ix": 5
+                    "ix": 5,
                   },
-                  "nm": "Transform"
-                }
+                  "nm": "Transform",
+                },
               ],
               "nm": "Combined Shape Copy",
               "np": 3,
               "cix": 2,
               "ix": 1,
               "mn": "ADBE Vector Group",
-              "hd": false
-            }
+              "hd": false,
+            },
           ],
           "ip": 0,
           "op": 848,
           "st": 0,
-          "bm": 0
-        }
-      ]
-    }
+          "bm": 0,
+        },
+      ],
+    },
   ],
   "layers": [
     {
@@ -332,25 +332,25 @@ export const isolatedStormsShowersDay = {
             {
               "i": {
                 "x": [0.833],
-                "y": [0.833]
+                "y": [0.833],
               },
               "o": {
                 "x": [0.167],
-                "y": [0.167]
+                "y": [0.167],
               },
               "n": ["0p833_0p833_0p167_0p167"],
               "t": 60,
               "s": [0],
-              "e": [100]
+              "e": [100],
             },
-            {"t": 106}
+            {"t": 106},
           ],
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 1,
@@ -358,57 +358,57 @@ export const isolatedStormsShowersDay = {
             {
               "i": {
                 "x": 0.36,
-                "y": 1
+                "y": 1,
               },
               "o": {
                 "x": 0.22,
-                "y": 1
+                "y": 1,
               },
               "n": "0p36_1_0p22_1",
               "t": 60,
               "s": [
                 1117,
                 969,
-                0
+                0,
               ],
               "e": [
                 1117,
                 1109,
-                0
+                0,
               ],
               "to": [
                 0,
                 23.3333339691162,
-                0
+                0,
               ],
               "ti": [
                 0,
                 -23.3333339691162,
-                0
-              ]
+                0,
+              ],
             },
-            {"t": 106}
+            {"t": 106},
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             105,
             105,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             90,
             90,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "w": 210,
@@ -416,7 +416,7 @@ export const isolatedStormsShowersDay = {
       "ip": 31,
       "op": 3631,
       "st": 31,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -432,25 +432,25 @@ export const isolatedStormsShowersDay = {
             {
               "i": {
                 "x": [0.833],
-                "y": [0.833]
+                "y": [0.833],
               },
               "o": {
                 "x": [0.167],
-                "y": [0.167]
+                "y": [0.167],
               },
               "n": ["0p833_0p833_0p167_0p167"],
               "t": 28,
               "s": [0],
-              "e": [100]
+              "e": [100],
             },
-            {"t": 74}
+            {"t": 74},
           ],
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 1,
@@ -458,57 +458,57 @@ export const isolatedStormsShowersDay = {
             {
               "i": {
                 "x": 0.36,
-                "y": 1
+                "y": 1,
               },
               "o": {
                 "x": 0.22,
-                "y": 1
+                "y": 1,
               },
               "n": "0p36_1_0p22_1",
               "t": 28,
               "s": [
                 1117,
                 505,
-                0
+                0,
               ],
               "e": [
                 1117,
                 645,
-                0
+                0,
               ],
               "to": [
                 0,
                 23.3333339691162,
-                0
+                0,
               ],
               "ti": [
                 0,
                 -23.3333339691162,
-                0
-              ]
+                0,
+              ],
             },
-            {"t": 74}
+            {"t": 74},
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             105,
             105,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             90,
             90,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "w": 210,
@@ -516,7 +516,7 @@ export const isolatedStormsShowersDay = {
       "ip": -1,
       "op": 3599,
       "st": -1,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -528,40 +528,40 @@ export const isolatedStormsShowersDay = {
         "o": {
           "a": 0,
           "k": 100,
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 0,
           "k": [
             880.5,
             1000,
-            0
+            0,
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             88,
             117,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             155,
             155,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "shapes": [
@@ -578,100 +578,100 @@ export const isolatedStormsShowersDay = {
                   "i": [
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
-                    ]
+                      0,
+                    ],
                   ],
                   "o": [
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
-                    ]
+                      0,
+                    ],
                   ],
                   "v": [
                     [
                       -29.333,
-                      -117
+                      -117,
                     ],
                     [
                       -88,
-                      0
+                      0,
                     ],
                     [
                       -35.2,
-                      0
+                      0,
                     ],
                     [
                       -41.741,
-                      117
+                      117,
                     ],
                     [
                       87.325,
-                      -29.25
+                      -29.25,
                     ],
                     [
                       41.052,
-                      -29.25
+                      -29.25,
                     ],
                     [
                       88,
-                      -117
-                    ]
+                      -117,
+                    ],
                   ],
-                  "c": true
+                  "c": true,
                 },
-                "ix": 2
+                "ix": 2,
               },
               "nm": "Path 2",
               "mn": "ADBE Vector Shape - Group",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "fl",
@@ -681,9 +681,9 @@ export const isolatedStormsShowersDay = {
                   1,
                   1,
                   1,
-                  1
+                  1,
                 ],
-                "ix": 4
+                "ix": 4,
               },
               "o": {
                 "a": 1,
@@ -691,39 +691,39 @@ export const isolatedStormsShowersDay = {
                   {
                     "i": {
                       "x": [0.833],
-                      "y": [0.833]
+                      "y": [0.833],
                     },
                     "o": {
                       "x": [0.167],
-                      "y": [0.167]
+                      "y": [0.167],
                     },
                     "n": ["0p833_0p833_0p167_0p167"],
                     "t": 46,
                     "s": [1],
-                    "e": [100]
+                    "e": [100],
                   },
                   {
                     "i": {
                       "x": [0.833],
-                      "y": [0.833]
+                      "y": [0.833],
                     },
                     "o": {
                       "x": [0.167],
-                      "y": [0.167]
+                      "y": [0.167],
                     },
                     "n": ["0p833_0p833_0p167_0p167"],
                     "t": 50,
                     "s": [100],
-                    "e": [3]
+                    "e": [3],
                   },
-                  {"t": 95}
+                  {"t": 95},
                 ],
-                "ix": 5
+                "ix": 5,
               },
               "r": 1,
               "nm": "Fill 1",
               "mn": "ADBE Vector Graphic - Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "tr",
@@ -731,61 +731,61 @@ export const isolatedStormsShowersDay = {
                 "a": 0,
                 "k": [
                   88,
-                  117
+                  117,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "a": {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 1
+                "ix": 1,
               },
               "s": {
                 "a": 0,
                 "k": [
                   100,
-                  100
+                  100,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "r": {
                 "a": 0,
                 "k": 0,
-                "ix": 6
+                "ix": 6,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 7
+                "ix": 7,
               },
               "sk": {
                 "a": 0,
                 "k": 0,
-                "ix": 4
+                "ix": 4,
               },
               "sa": {
                 "a": 0,
                 "k": 0,
-                "ix": 5
+                "ix": 5,
               },
-              "nm": "Transform"
-            }
+              "nm": "Transform",
+            },
           ],
           "nm": "Group 1",
           "np": 2,
           "cix": 2,
           "ix": 1,
           "mn": "ADBE Vector Group",
-          "hd": false
-        }
+          "hd": false,
+        },
       ],
       "ip": -9,
       "op": 839,
       "st": -9,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -800,53 +800,53 @@ export const isolatedStormsShowersDay = {
             {
               "i": {
                 "x": [0.833],
-                "y": [0.833]
+                "y": [0.833],
               },
               "o": {
                 "x": [0.167],
-                "y": [0.167]
+                "y": [0.167],
               },
               "n": ["0p833_0p833_0p167_0p167"],
               "t": 50,
               "s": [0],
-              "e": [100]
+              "e": [100],
             },
-            {"t": 57}
+            {"t": 57},
           ],
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 0,
           "k": [
             880.5,
             1000,
-            0
+            0,
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             88,
             117,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             155,
             155,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "shapes": [
@@ -863,100 +863,100 @@ export const isolatedStormsShowersDay = {
                   "i": [
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
-                    ]
+                      0,
+                    ],
                   ],
                   "o": [
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      0
-                    ]
+                      0,
+                    ],
                   ],
                   "v": [
                     [
                       -29.333,
-                      -117
+                      -117,
                     ],
                     [
                       -88,
-                      0
+                      0,
                     ],
                     [
                       -35.2,
-                      0
+                      0,
                     ],
                     [
                       -41.741,
-                      117
+                      117,
                     ],
                     [
                       87.325,
-                      -29.25
+                      -29.25,
                     ],
                     [
                       41.052,
-                      -29.25
+                      -29.25,
                     ],
                     [
                       88,
-                      -117
-                    ]
+                      -117,
+                    ],
                   ],
-                  "c": true
+                  "c": true,
                 },
-                "ix": 2
+                "ix": 2,
               },
               "nm": "Path 2",
               "mn": "ADBE Vector Shape - Group",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "fl",
@@ -966,19 +966,19 @@ export const isolatedStormsShowersDay = {
                   0.98431372549,
                   0.737254901961,
                   0.01568627451,
-                  1
+                  1,
                 ],
-                "ix": 4
+                "ix": 4,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 5
+                "ix": 5,
               },
               "r": 1,
               "nm": "Fill 1",
               "mn": "ADBE Vector Graphic - Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "tr",
@@ -986,61 +986,61 @@ export const isolatedStormsShowersDay = {
                 "a": 0,
                 "k": [
                   88,
-                  117
+                  117,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "a": {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 1
+                "ix": 1,
               },
               "s": {
                 "a": 0,
                 "k": [
                   100,
-                  100
+                  100,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "r": {
                 "a": 0,
                 "k": 0,
-                "ix": 6
+                "ix": 6,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 7
+                "ix": 7,
               },
               "sk": {
                 "a": 0,
                 "k": 0,
-                "ix": 4
+                "ix": 4,
               },
               "sa": {
                 "a": 0,
                 "k": 0,
-                "ix": 5
+                "ix": 5,
               },
-              "nm": "Transform"
-            }
+              "nm": "Transform",
+            },
           ],
           "nm": "Group 1",
           "np": 2,
           "cix": 2,
           "ix": 1,
           "mn": "ADBE Vector Group",
-          "hd": false
-        }
+          "hd": false,
+        },
       ],
       "ip": -9,
       "op": 839,
       "st": -9,
-      "bm": 1
+      "bm": 1,
     },
     {
       "ddd": 0,
@@ -1052,12 +1052,12 @@ export const isolatedStormsShowersDay = {
         "o": {
           "a": 0,
           "k": 0,
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 1,
@@ -1065,47 +1065,47 @@ export const isolatedStormsShowersDay = {
             {
               "i": {
                 "x": 0.36,
-                "y": 1
+                "y": 1,
               },
               "o": {
                 "x": 0.22,
-                "y": 1
+                "y": 1,
               },
               "n": "0p36_1_0p22_1",
               "t": 1,
               "s": [
                 422,
                 750,
-                0
+                0,
               ],
               "e": [
                 750,
                 750,
-                0
+                0,
               ],
               "to": [
                 54.6666679382324,
                 0,
-                0
+                0,
               ],
               "ti": [
                 -54.6666679382324,
                 0,
-                0
-              ]
+                0,
+              ],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             0,
             0,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 1,
@@ -1115,53 +1115,53 @@ export const isolatedStormsShowersDay = {
                 "x": [
                   0.36,
                   0.36,
-                  0.36
+                  0.36,
                 ],
                 "y": [
                   1,
                   1,
-                  1
-                ]
+                  1,
+                ],
               },
               "o": {
                 "x": [
                   0.22,
                   0.22,
-                  0.22
+                  0.22,
                 ],
                 "y": [
                   0,
                   0,
-                  0
-                ]
+                  0,
+                ],
               },
               "n": [
                 "0p36_1_0p22_0",
                 "0p36_1_0p22_0",
-                "0p36_1_0p22_0"
+                "0p36_1_0p22_0",
               ],
               "t": 1,
               "s": [
                 100,
                 100,
-                100
+                100,
               ],
               "e": [
                 100,
                 100,
-                100
-              ]
+                100,
+              ],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "ip": -33,
       "op": 849,
       "st": -2,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -1177,53 +1177,53 @@ export const isolatedStormsShowersDay = {
             {
               "i": {
                 "x": [0.518],
-                "y": [0.936]
+                "y": [0.936],
               },
               "o": {
                 "x": [0.294],
-                "y": [0.52]
+                "y": [0.52],
               },
               "n": ["0p518_0p936_0p294_0p52"],
               "t": 1,
               "s": [0],
-              "e": [100]
+              "e": [100],
             },
-            {"t": 37}
+            {"t": 37},
           ],
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 0,
           "k": [
             -250,
             -500,
-            0
+            0,
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             0,
             0,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             100,
             100,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "shapes": [
@@ -1237,21 +1237,21 @@ export const isolatedStormsShowersDay = {
                 "a": 0,
                 "k": [
                   24,
-                  24
+                  24,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "p": {
                 "a": 0,
                 "k": [
                   12,
-                  12
+                  12,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "nm": "Ellipse Path 1",
               "mn": "ADBE Vector Shape - Ellipse",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "fl",
@@ -1261,26 +1261,26 @@ export const isolatedStormsShowersDay = {
                   0.945097982883,
                   0.952941000462,
                   0.956862986088,
-                  1
+                  1,
                 ],
-                "ix": 4
+                "ix": 4,
               },
               "o": {
                 "a": 0,
                 "k": 80,
-                "ix": 5
+                "ix": 5,
               },
               "r": 2,
               "nm": "Fill 1",
               "mn": "ADBE Vector Graphic - Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "gf",
               "o": {
                 "a": 0,
                 "k": 50,
-                "ix": 10
+                "ix": 10,
               },
               "r": 2,
               "g": {
@@ -1305,31 +1305,31 @@ export const isolatedStormsShowersDay = {
                     0.5,
                     0.1,
                     1,
-                    0.2
+                    0.2,
                   ],
-                  "ix": 9
-                }
+                  "ix": 9,
+                },
               },
               "s": {
                 "a": 0,
                 "k": [
                   11.385,
-                  13.566
+                  13.566,
                 ],
-                "ix": 5
+                "ix": 5,
               },
               "e": {
                 "a": 0,
                 "k": [
                   23.401,
-                  24.13
+                  24.13,
                 ],
-                "ix": 6
+                "ix": 6,
               },
               "t": 1,
               "nm": "Gradient Fill 1",
               "mn": "ADBE Vector Graphic - G-Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "tr",
@@ -1337,61 +1337,61 @@ export const isolatedStormsShowersDay = {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "a": {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 1
+                "ix": 1,
               },
               "s": {
                 "a": 0,
                 "k": [
                   3125,
-                  3125
+                  3125,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "r": {
                 "a": 0,
                 "k": 0,
-                "ix": 6
+                "ix": 6,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 7
+                "ix": 7,
               },
               "sk": {
                 "a": 0,
                 "k": 0,
-                "ix": 4
+                "ix": 4,
               },
               "sa": {
                 "a": 0,
                 "k": 0,
-                "ix": 5
+                "ix": 5,
               },
-              "nm": "Transform"
-            }
+              "nm": "Transform",
+            },
           ],
           "nm": "Oval",
           "np": 3,
           "cix": 2,
           "ix": 1,
           "mn": "ADBE Vector Group",
-          "hd": false
-        }
+          "hd": false,
+        },
       ],
       "ip": -58,
       "op": 842,
       "st": -6,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -1406,25 +1406,25 @@ export const isolatedStormsShowersDay = {
             {
               "i": {
                 "x": [0.36],
-                "y": [1]
+                "y": [1],
               },
               "o": {
                 "x": [0.22],
-                "y": [99]
+                "y": [99],
               },
               "n": ["0p36_1_0p22_99"],
               "t": 1,
               "s": [100],
-              "e": [100]
+              "e": [100],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 1,
@@ -1432,57 +1432,57 @@ export const isolatedStormsShowersDay = {
             {
               "i": {
                 "x": 0.36,
-                "y": 1
+                "y": 1,
               },
               "o": {
                 "x": 0.22,
-                "y": 1
+                "y": 1,
               },
               "n": "0p36_1_0p22_1",
               "t": 1,
               "s": [
                 487.75,
                 237,
-                0
+                0,
               ],
               "e": [
                 232.75,
                 237,
-                0
+                0,
               ],
               "to": [
                 -42.5,
                 0,
-                0
+                0,
               ],
               "ti": [
                 42.5,
                 0,
-                0
-              ]
+                0,
+              ],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             -496,
             -494,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             100,
             100,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "shapes": [
@@ -1496,21 +1496,21 @@ export const isolatedStormsShowersDay = {
                 "a": 0,
                 "k": [
                   365.547,
-                  365.547
+                  365.547,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "p": {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "nm": "Ellipse Path 1",
               "mn": "ADBE Vector Shape - Ellipse",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "tr",
@@ -1518,55 +1518,55 @@ export const isolatedStormsShowersDay = {
                 "a": 0,
                 "k": [
                   -499.359,
-                  -499.43
+                  -499.43,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "a": {
                 "a": 0,
                 "k": [
                   -209.552,
-                  -208.357
+                  -208.357,
                 ],
-                "ix": 1
+                "ix": 1,
               },
               "s": {
                 "a": 0,
                 "k": [
                   131.101,
-                  128.967
+                  128.967,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "r": {
                 "a": 0,
                 "k": 0,
-                "ix": 6
+                "ix": 6,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 7
+                "ix": 7,
               },
               "sk": {
                 "a": 0,
                 "k": 0,
-                "ix": 4
+                "ix": 4,
               },
               "sa": {
                 "a": 0,
                 "k": 0,
-                "ix": 5
+                "ix": 5,
               },
-              "nm": "Transform"
-            }
+              "nm": "Transform",
+            },
           ],
           "nm": "Ellipse 1",
           "np": 1,
           "cix": 2,
           "ix": 1,
           "mn": "ADBE Vector Group",
-          "hd": false
+          "hd": false,
         },
         {
           "ty": "st",
@@ -1576,14 +1576,14 @@ export const isolatedStormsShowersDay = {
               0.98431372549,
               0.737254901961,
               0.01568627451,
-              1
+              1,
             ],
-            "ix": 3
+            "ix": 3,
           },
           "o": {
             "a": 0,
             "k": 100,
-            "ix": 4
+            "ix": 4,
           },
           "w": {
             "a": 1,
@@ -1591,33 +1591,33 @@ export const isolatedStormsShowersDay = {
               {
                 "i": {
                   "x": [0.36],
-                  "y": [1]
+                  "y": [1],
                 },
                 "o": {
                   "x": [0.22],
-                  "y": [1]
+                  "y": [1],
                 },
                 "n": ["0p36_1_0p22_1"],
                 "t": 1,
                 "s": [1],
-                "e": [30]
+                "e": [30],
               },
-              {"t": 75}
+              {"t": 75},
             ],
-            "ix": 5
+            "ix": 5,
           },
           "lc": 1,
           "lj": 1,
           "ml": 4,
           "nm": "Stroke 1",
           "mn": "ADBE Vector Graphic - Stroke",
-          "hd": false
-        }
+          "hd": false,
+        },
       ],
       "ip": 1,
       "op": 901,
       "st": -1,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -1631,40 +1631,40 @@ export const isolatedStormsShowersDay = {
         "o": {
           "a": 0,
           "k": 100,
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 0,
           "k": [
             -250,
             -500,
-            0
+            0,
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             0,
             0,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             100,
             100,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "shapes": [
@@ -1678,21 +1678,21 @@ export const isolatedStormsShowersDay = {
                 "a": 0,
                 "k": [
                   24,
-                  24
+                  24,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "p": {
                 "a": 0,
                 "k": [
                   12,
-                  12
+                  12,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "nm": "Ellipse Path 1",
               "mn": "ADBE Vector Shape - Ellipse",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "fl",
@@ -1702,26 +1702,26 @@ export const isolatedStormsShowersDay = {
                   0.945097982883,
                   0.952941000462,
                   0.956862986088,
-                  1
+                  1,
                 ],
-                "ix": 4
+                "ix": 4,
               },
               "o": {
                 "a": 0,
                 "k": 80,
-                "ix": 5
+                "ix": 5,
               },
               "r": 2,
               "nm": "Fill 1",
               "mn": "ADBE Vector Graphic - Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "gf",
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 10
+                "ix": 10,
               },
               "r": 2,
               "g": {
@@ -1746,31 +1746,31 @@ export const isolatedStormsShowersDay = {
                     0.5,
                     0.1,
                     1,
-                    0.2
+                    0.2,
                   ],
-                  "ix": 9
-                }
+                  "ix": 9,
+                },
               },
               "s": {
                 "a": 0,
                 "k": [
                   11.385,
-                  13.566
+                  13.566,
                 ],
-                "ix": 5
+                "ix": 5,
               },
               "e": {
                 "a": 0,
                 "k": [
                   23.401,
-                  24.13
+                  24.13,
                 ],
-                "ix": 6
+                "ix": 6,
               },
               "t": 1,
               "nm": "Gradient Fill 1",
               "mn": "ADBE Vector Graphic - G-Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "tr",
@@ -1778,61 +1778,61 @@ export const isolatedStormsShowersDay = {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "a": {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 1
+                "ix": 1,
               },
               "s": {
                 "a": 0,
                 "k": [
                   3125,
-                  3125
+                  3125,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "r": {
                 "a": 0,
                 "k": 0,
-                "ix": 6
+                "ix": 6,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 7
+                "ix": 7,
               },
               "sk": {
                 "a": 0,
                 "k": 0,
-                "ix": 4
+                "ix": 4,
               },
               "sa": {
                 "a": 0,
                 "k": 0,
-                "ix": 5
+                "ix": 5,
               },
-              "nm": "Transform"
-            }
+              "nm": "Transform",
+            },
           ],
           "nm": "Oval",
           "np": 3,
           "cix": 2,
           "ix": 1,
           "mn": "ADBE Vector Group",
-          "hd": false
-        }
+          "hd": false,
+        },
       ],
       "ip": -58,
       "op": 842,
       "st": -6,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -1849,53 +1849,53 @@ export const isolatedStormsShowersDay = {
             {
               "i": {
                 "x": [0.36],
-                "y": [1]
+                "y": [1],
               },
               "o": {
                 "x": [0.22],
-                "y": [1]
+                "y": [1],
               },
               "n": ["0p36_1_0p22_1"],
               "t": 1,
               "s": [0],
-              "e": [100]
+              "e": [100],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 0,
           "k": [
             -84,
             -62.5,
-            0
+            0,
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             416,
             0,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             100,
             100,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "shapes": [
@@ -1912,52 +1912,52 @@ export const isolatedStormsShowersDay = {
                   "i": [
                     [
                       5.52,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      -5.52
+                      -5.52,
                     ],
                     [
                       0.01,
-                      -0.02
-                    ]
+                      -0.02,
+                    ],
                   ],
                   "o": [
                     [
                       -5.52,
-                      0
+                      0,
                     ],
                     [
                       0.01,
-                      -0.02
+                      -0.02,
                     ],
                     [
                       0,
-                      -5.52
-                    ]
+                      -5.52,
+                    ],
                   ],
                   "v": [
                     [
                       10,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      10
+                      10,
                     ],
                     [
                       20,
-                      10
-                    ]
+                      10,
+                    ],
                   ],
-                  "c": true
+                  "c": true,
                 },
-                "ix": 2
+                "ix": 2,
               },
               "nm": "Path 1",
               "mn": "ADBE Vector Shape - Group",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "fl",
@@ -1967,19 +1967,19 @@ export const isolatedStormsShowersDay = {
                   0.945097982883,
                   0.952941000462,
                   0.956862986088,
-                  1
+                  1,
                 ],
-                "ix": 4
+                "ix": 4,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 5
+                "ix": 5,
               },
               "r": 1,
               "nm": "Fill 1",
               "mn": "ADBE Vector Graphic - Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "tr",
@@ -1987,62 +1987,62 @@ export const isolatedStormsShowersDay = {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "a": {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 1
+                "ix": 1,
               },
               "s": {
                 "a": 0,
                 "k": [
                   3125,
-                  3125
+                  3125,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "r": {
                 "a": 0,
                 "k": 0,
-                "ix": 6
+                "ix": 6,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 7
+                "ix": 7,
               },
               "sk": {
                 "a": 0,
                 "k": 0,
-                "ix": 4
+                "ix": 4,
               },
               "sa": {
                 "a": 0,
                 "k": 0,
-                "ix": 5
+                "ix": 5,
               },
-              "nm": "Transform"
-            }
+              "nm": "Transform",
+            },
           ],
           "nm": "Group 1",
           "np": 2,
           "cix": 2,
           "ix": 1,
           "mn": "ADBE Vector Group",
-          "hd": false
-        }
+          "hd": false,
+        },
       ],
       "ip": -52,
       "op": 848,
       "st": -70,
-      "bm": 0
-    }
+      "bm": 0,
+    },
   ],
-  "markers": []
+  "markers": [],
 };
