@@ -21,40 +21,40 @@ export const mixedSleet = {
             "o": {
               "a": 0,
               "k": 100,
-              "ix": 11
+              "ix": 11,
             },
             "r": {
               "a": 0,
               "k": -45,
-              "ix": 10
+              "ix": 10,
             },
             "p": {
               "a": 0,
               "k": [
                 104.5,
                 105.5,
-                0
+                0,
               ],
-              "ix": 2
+              "ix": 2,
             },
             "a": {
               "a": 0,
               "k": [
                 7.953,
                 7.953,
-                0
+                0,
               ],
-              "ix": 1
+              "ix": 1,
             },
             "s": {
               "a": 0,
               "k": [
                 1031.5,
                 1031.5,
-                100
+                100,
               ],
-              "ix": 6
-            }
+              "ix": 6,
+            },
           },
           "ao": 0,
           "shapes": [
@@ -71,21 +71,21 @@ export const mixedSleet = {
                         "a": 0,
                         "k": [
                           16,
-                          16
+                          16,
                         ],
-                        "ix": 2
+                        "ix": 2,
                       },
                       "p": {
                         "a": 0,
                         "k": [
                           8,
-                          8
+                          8,
                         ],
-                        "ix": 3
+                        "ix": 3,
                       },
                       "nm": "Ellipse Path 1",
                       "mn": "ADBE Vector Shape - Ellipse",
-                      "hd": false
+                      "hd": false,
                     },
                     {
                       "ty": "tr",
@@ -93,55 +93,55 @@ export const mixedSleet = {
                         "a": 0,
                         "k": [
                           8,
-                          8
+                          8,
                         ],
-                        "ix": 2
+                        "ix": 2,
                       },
                       "a": {
                         "a": 0,
                         "k": [
                           8,
-                          8
+                          8,
                         ],
-                        "ix": 1
+                        "ix": 1,
                       },
                       "s": {
                         "a": 0,
                         "k": [
                           100,
-                          100
+                          100,
                         ],
-                        "ix": 3
+                        "ix": 3,
                       },
                       "r": {
                         "a": 0,
                         "k": 0,
-                        "ix": 6
+                        "ix": 6,
                       },
                       "o": {
                         "a": 0,
                         "k": 100,
-                        "ix": 7
+                        "ix": 7,
                       },
                       "sk": {
                         "a": 0,
                         "k": 0,
-                        "ix": 4
+                        "ix": 4,
                       },
                       "sa": {
                         "a": 0,
                         "k": 0,
-                        "ix": 5
+                        "ix": 5,
                       },
-                      "nm": "Transform"
-                    }
+                      "nm": "Transform",
+                    },
                   ],
                   "nm": "Oval",
                   "np": 1,
                   "cix": 2,
                   "ix": 1,
                   "mn": "ADBE Vector Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ty": "tr",
@@ -149,55 +149,55 @@ export const mixedSleet = {
                     "a": 0,
                     "k": [
                       0,
-                      0
+                      0,
                     ],
-                    "ix": 2
+                    "ix": 2,
                   },
                   "a": {
                     "a": 0,
                     "k": [
                       0,
-                      0
+                      0,
                     ],
-                    "ix": 1
+                    "ix": 1,
                   },
                   "s": {
                     "a": 0,
                     "k": [
                       100,
-                      100
+                      100,
                     ],
-                    "ix": 3
+                    "ix": 3,
                   },
                   "r": {
                     "a": 0,
                     "k": 0,
-                    "ix": 6
+                    "ix": 6,
                   },
                   "o": {
                     "a": 0,
                     "k": 100,
-                    "ix": 7
+                    "ix": 7,
                   },
                   "sk": {
                     "a": 0,
                     "k": 0,
-                    "ix": 4
+                    "ix": 4,
                   },
                   "sa": {
                     "a": 0,
                     "k": 0,
-                    "ix": 5
+                    "ix": 5,
                   },
-                  "nm": "Transform"
-                }
+                  "nm": "Transform",
+                },
               ],
               "nm": "Combined Shape Copy",
               "np": 2,
               "cix": 2,
               "ix": 1,
               "mn": "ADBE Vector Group",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "st",
@@ -207,34 +207,34 @@ export const mixedSleet = {
                   0.258823529412,
                   0.521568627451,
                   0.956862745098,
-                  1
+                  1,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 4
+                "ix": 4,
               },
               "w": {
                 "a": 0,
                 "k": 3,
-                "ix": 5
+                "ix": 5,
               },
               "lc": 1,
               "lj": 1,
               "ml": 4,
               "nm": "Stroke 1",
               "mn": "ADBE Vector Graphic - Stroke",
-              "hd": false
-            }
+              "hd": false,
+            },
           ],
           "ip": 0,
           "op": 848,
           "st": 0,
-          "bm": 0
-        }
-      ]
+          "bm": 0,
+        },
+      ],
     },
     {
       "id": "comp_1070",
@@ -249,40 +249,40 @@ export const mixedSleet = {
             "o": {
               "a": 0,
               "k": 100,
-              "ix": 11
+              "ix": 11,
             },
             "r": {
               "a": 0,
               "k": 0,
-              "ix": 10
+              "ix": 10,
             },
             "p": {
               "a": 0,
               "k": [
                 105,
                 105,
-                0
+                0,
               ],
-              "ix": 2
+              "ix": 2,
             },
             "a": {
               "a": 0,
               "k": [
                 6,
                 4.5,
-                0
+                0,
               ],
-              "ix": 1
+              "ix": 1,
             },
             "s": {
               "a": 0,
               "k": [
                 1887,
                 1887,
-                100
+                100,
               ],
-              "ix": 6
-            }
+              "ix": 6,
+            },
           },
           "ao": 0,
           "shapes": [
@@ -299,40 +299,40 @@ export const mixedSleet = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           6,
-                          9
+                          9,
                         ],
                         [
                           6,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 1",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ind": 1,
@@ -344,40 +344,40 @@ export const mixedSleet = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           1.5,
-                          9
+                          9,
                         ],
                         [
                           1.5,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 2",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ind": 2,
@@ -389,40 +389,40 @@ export const mixedSleet = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           10.5,
-                          9
+                          9,
                         ],
                         [
                           10.5,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 3",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ty": "st",
@@ -432,26 +432,26 @@ export const mixedSleet = {
                       0.258823529412,
                       0.521568627451,
                       0.956862745098,
-                      1
+                      1,
                     ],
-                    "ix": 3
+                    "ix": 3,
                   },
                   "o": {
                     "a": 0,
                     "k": 100,
-                    "ix": 4
+                    "ix": 4,
                   },
                   "w": {
                     "a": 0,
                     "k": 1.5,
-                    "ix": 5
+                    "ix": 5,
                   },
                   "lc": 1,
                   "lj": 1,
                   "ml": 10,
                   "nm": "Stroke 1",
                   "mn": "ADBE Vector Graphic - Stroke",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ty": "tr",
@@ -459,63 +459,63 @@ export const mixedSleet = {
                     "a": 0,
                     "k": [
                       0,
-                      0
+                      0,
                     ],
-                    "ix": 2
+                    "ix": 2,
                   },
                   "a": {
                     "a": 0,
                     "k": [
                       0,
-                      0
+                      0,
                     ],
-                    "ix": 1
+                    "ix": 1,
                   },
                   "s": {
                     "a": 0,
                     "k": [
                       100,
-                      100
+                      100,
                     ],
-                    "ix": 3
+                    "ix": 3,
                   },
                   "r": {
                     "a": 0,
                     "k": 0,
-                    "ix": 6
+                    "ix": 6,
                   },
                   "o": {
                     "a": 0,
                     "k": 100,
-                    "ix": 7
+                    "ix": 7,
                   },
                   "sk": {
                     "a": 0,
                     "k": 0,
-                    "ix": 4
+                    "ix": 4,
                   },
                   "sa": {
                     "a": 0,
                     "k": 0,
-                    "ix": 5
+                    "ix": 5,
                   },
-                  "nm": "Transform"
-                }
+                  "nm": "Transform",
+                },
               ],
               "nm": "Group 1",
               "np": 4,
               "cix": 2,
               "ix": 1,
               "mn": "ADBE Vector Group",
-              "hd": false
-            }
+              "hd": false,
+            },
           ],
           "ip": 0,
           "op": 3600,
           "st": 0,
-          "bm": 0
-        }
-      ]
+          "bm": 0,
+        },
+      ],
     },
     {
       "id": "comp_1071",
@@ -530,40 +530,40 @@ export const mixedSleet = {
             "o": {
               "a": 0,
               "k": 100,
-              "ix": 11
+              "ix": 11,
             },
             "r": {
               "a": 0,
               "k": -45,
-              "ix": 10
+              "ix": 10,
             },
             "p": {
               "a": 0,
               "k": [
                 -15,
                 122.5,
-                0
+                0,
               ],
-              "ix": 2
+              "ix": 2,
             },
             "a": {
               "a": 0,
               "k": [
                 0,
                 0,
-                0
+                0,
               ],
-              "ix": 1
+              "ix": 1,
             },
             "s": {
               "a": 0,
               "k": [
                 1062.5,
                 1062.5,
-                100
+                100,
               ],
-              "ix": 6
-            }
+              "ix": 6,
+            },
           },
           "ao": 0,
           "shapes": [
@@ -580,21 +580,21 @@ export const mixedSleet = {
                         "a": 0,
                         "k": [
                           16,
-                          16
+                          16,
                         ],
-                        "ix": 2
+                        "ix": 2,
                       },
                       "p": {
                         "a": 0,
                         "k": [
                           8,
-                          8
+                          8,
                         ],
-                        "ix": 3
+                        "ix": 3,
                       },
                       "nm": "Ellipse Path 1",
                       "mn": "ADBE Vector Shape - Ellipse",
-                      "hd": false
+                      "hd": false,
                     },
                     {
                       "ty": "tr",
@@ -602,55 +602,55 @@ export const mixedSleet = {
                         "a": 0,
                         "k": [
                           8,
-                          8
+                          8,
                         ],
-                        "ix": 2
+                        "ix": 2,
                       },
                       "a": {
                         "a": 0,
                         "k": [
                           8,
-                          8
+                          8,
                         ],
-                        "ix": 1
+                        "ix": 1,
                       },
                       "s": {
                         "a": 0,
                         "k": [
                           100,
-                          100
+                          100,
                         ],
-                        "ix": 3
+                        "ix": 3,
                       },
                       "r": {
                         "a": 0,
                         "k": 0,
-                        "ix": 6
+                        "ix": 6,
                       },
                       "o": {
                         "a": 0,
                         "k": 100,
-                        "ix": 7
+                        "ix": 7,
                       },
                       "sk": {
                         "a": 0,
                         "k": 0,
-                        "ix": 4
+                        "ix": 4,
                       },
                       "sa": {
                         "a": 0,
                         "k": 0,
-                        "ix": 5
+                        "ix": 5,
                       },
-                      "nm": "Transform"
-                    }
+                      "nm": "Transform",
+                    },
                   ],
                   "nm": "Oval",
                   "np": 1,
                   "cix": 2,
                   "ix": 1,
                   "mn": "ADBE Vector Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ty": "gr",
@@ -662,26 +662,26 @@ export const mixedSleet = {
                         "a": 0,
                         "k": [
                           8,
-                          8
+                          8,
                         ],
-                        "ix": 2
+                        "ix": 2,
                       },
                       "p": {
                         "a": 0,
                         "k": [
                           12,
-                          4
+                          4,
                         ],
-                        "ix": 3
+                        "ix": 3,
                       },
                       "r": {
                         "a": 0,
                         "k": 0,
-                        "ix": 4
+                        "ix": 4,
                       },
                       "nm": "Rectangle Path 1",
                       "mn": "ADBE Vector Shape - Rect",
-                      "hd": false
+                      "hd": false,
                     },
                     {
                       "ty": "tr",
@@ -689,55 +689,55 @@ export const mixedSleet = {
                         "a": 0,
                         "k": [
                           12,
-                          4
+                          4,
                         ],
-                        "ix": 2
+                        "ix": 2,
                       },
                       "a": {
                         "a": 0,
                         "k": [
                           12,
-                          4
+                          4,
                         ],
-                        "ix": 1
+                        "ix": 1,
                       },
                       "s": {
                         "a": 0,
                         "k": [
                           100,
-                          100
+                          100,
                         ],
-                        "ix": 3
+                        "ix": 3,
                       },
                       "r": {
                         "a": 0,
                         "k": 0,
-                        "ix": 6
+                        "ix": 6,
                       },
                       "o": {
                         "a": 0,
                         "k": 100,
-                        "ix": 7
+                        "ix": 7,
                       },
                       "sk": {
                         "a": 0,
                         "k": 0,
-                        "ix": 4
+                        "ix": 4,
                       },
                       "sa": {
                         "a": 0,
                         "k": 0,
-                        "ix": 5
+                        "ix": 5,
                       },
-                      "nm": "Transform"
-                    }
+                      "nm": "Transform",
+                    },
                   ],
                   "nm": "Rectangle 2",
                   "np": 1,
                   "cix": 2,
                   "ix": 2,
                   "mn": "ADBE Vector Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ty": "fl",
@@ -747,19 +747,19 @@ export const mixedSleet = {
                       0.258823990822,
                       0.521569013596,
                       0.956862986088,
-                      1
+                      1,
                     ],
-                    "ix": 4
+                    "ix": 4,
                   },
                   "o": {
                     "a": 0,
                     "k": 100,
-                    "ix": 5
+                    "ix": 5,
                   },
                   "r": 1,
                   "nm": "Fill 1",
                   "mn": "ADBE Vector Graphic - Fill",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ty": "tr",
@@ -767,64 +767,64 @@ export const mixedSleet = {
                     "a": 0,
                     "k": [
                       0,
-                      0
+                      0,
                     ],
-                    "ix": 2
+                    "ix": 2,
                   },
                   "a": {
                     "a": 0,
                     "k": [
                       0,
-                      0
+                      0,
                     ],
-                    "ix": 1
+                    "ix": 1,
                   },
                   "s": {
                     "a": 0,
                     "k": [
                       100,
-                      100
+                      100,
                     ],
-                    "ix": 3
+                    "ix": 3,
                   },
                   "r": {
                     "a": 0,
                     "k": 0,
-                    "ix": 6
+                    "ix": 6,
                   },
                   "o": {
                     "a": 0,
                     "k": 100,
-                    "ix": 7
+                    "ix": 7,
                   },
                   "sk": {
                     "a": 0,
                     "k": 0,
-                    "ix": 4
+                    "ix": 4,
                   },
                   "sa": {
                     "a": 0,
                     "k": 0,
-                    "ix": 5
+                    "ix": 5,
                   },
-                  "nm": "Transform"
-                }
+                  "nm": "Transform",
+                },
               ],
               "nm": "Combined Shape Copy",
               "np": 3,
               "cix": 2,
               "ix": 1,
               "mn": "ADBE Vector Group",
-              "hd": false
-            }
+              "hd": false,
+            },
           ],
           "ip": 0,
           "op": 848,
           "st": 0,
-          "bm": 0
-        }
-      ]
-    }
+          "bm": 0,
+        },
+      ],
+    },
   ],
   "layers": [
     {
@@ -841,25 +841,25 @@ export const mixedSleet = {
             {
               "i": {
                 "x": [0.833],
-                "y": [0.833]
+                "y": [0.833],
               },
               "o": {
                 "x": [0.167],
-                "y": [0.167]
+                "y": [0.167],
               },
               "n": ["0p833_0p833_0p167_0p167"],
               "t": 60,
               "s": [0],
-              "e": [100]
+              "e": [100],
             },
-            {"t": 106}
+            {"t": 106},
           ],
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 1,
@@ -867,57 +867,57 @@ export const mixedSleet = {
             {
               "i": {
                 "x": 0.36,
-                "y": 1
+                "y": 1,
               },
               "o": {
                 "x": 0.22,
-                "y": 1
+                "y": 1,
               },
               "n": "0p36_1_0p22_1",
               "t": 60,
               "s": [
                 1117,
                 969,
-                0
+                0,
               ],
               "e": [
                 1117,
                 1109,
-                0
+                0,
               ],
               "to": [
                 0,
                 23.3333339691162,
-                0
+                0,
               ],
               "ti": [
                 0,
                 -23.3333339691162,
-                0
-              ]
+                0,
+              ],
             },
-            {"t": 106}
+            {"t": 106},
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             105,
             105,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             90,
             90,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "w": 210,
@@ -925,7 +925,7 @@ export const mixedSleet = {
       "ip": 31,
       "op": 3631,
       "st": 31,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -941,25 +941,25 @@ export const mixedSleet = {
             {
               "i": {
                 "x": [0.833],
-                "y": [0.833]
+                "y": [0.833],
               },
               "o": {
                 "x": [0.167],
-                "y": [0.167]
+                "y": [0.167],
               },
               "n": ["0p833_0p833_0p167_0p167"],
               "t": 44,
               "s": [0],
-              "e": [100]
+              "e": [100],
             },
-            {"t": 90}
+            {"t": 90},
           ],
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 1,
@@ -967,57 +967,57 @@ export const mixedSleet = {
             {
               "i": {
                 "x": 0.36,
-                "y": 1
+                "y": 1,
               },
               "o": {
                 "x": 0.22,
-                "y": 1
+                "y": 1,
               },
               "n": "0p36_1_0p22_1",
               "t": 44,
               "s": [
                 861,
                 737,
-                0
+                0,
               ],
               "e": [
                 861,
                 877,
-                0
+                0,
               ],
               "to": [
                 0,
                 23.3333339691162,
-                0
+                0,
               ],
               "ti": [
                 0,
                 -23.3333339691162,
-                0
-              ]
+                0,
+              ],
             },
-            {"t": 90}
+            {"t": 90},
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             105,
             105,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             90,
             90,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "w": 210,
@@ -1025,7 +1025,7 @@ export const mixedSleet = {
       "ip": 15,
       "op": 3615,
       "st": 15,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -1041,25 +1041,25 @@ export const mixedSleet = {
             {
               "i": {
                 "x": [0.833],
-                "y": [0.833]
+                "y": [0.833],
               },
               "o": {
                 "x": [0.167],
-                "y": [0.167]
+                "y": [0.167],
               },
               "n": ["0p833_0p833_0p167_0p167"],
               "t": 28,
               "s": [0],
-              "e": [100]
+              "e": [100],
             },
-            {"t": 74}
+            {"t": 74},
           ],
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 1,
@@ -1067,57 +1067,57 @@ export const mixedSleet = {
             {
               "i": {
                 "x": 0.36,
-                "y": 1
+                "y": 1,
               },
               "o": {
                 "x": 0.22,
-                "y": 1
+                "y": 1,
               },
               "n": "0p36_1_0p22_1",
               "t": 28,
               "s": [
                 1117,
                 505,
-                0
+                0,
               ],
               "e": [
                 1117,
                 645,
-                0
+                0,
               ],
               "to": [
                 0,
                 23.3333339691162,
-                0
+                0,
               ],
               "ti": [
                 0,
                 -23.3333339691162,
-                0
-              ]
+                0,
+              ],
             },
-            {"t": 74}
+            {"t": 74},
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             105,
             105,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             90,
             90,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "w": 210,
@@ -1125,7 +1125,7 @@ export const mixedSleet = {
       "ip": -1,
       "op": 3599,
       "st": -1,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -1137,12 +1137,12 @@ export const mixedSleet = {
         "o": {
           "a": 0,
           "k": 0,
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 1,
@@ -1150,47 +1150,47 @@ export const mixedSleet = {
             {
               "i": {
                 "x": 0.36,
-                "y": 1
+                "y": 1,
               },
               "o": {
                 "x": 0.22,
-                "y": 1
+                "y": 1,
               },
               "n": "0p36_1_0p22_1",
               "t": 1,
               "s": [
                 422,
                 750,
-                0
+                0,
               ],
               "e": [
                 750,
                 750,
-                0
+                0,
               ],
               "to": [
                 54.6666679382324,
                 0,
-                0
+                0,
               ],
               "ti": [
                 -54.6666679382324,
                 0,
-                0
-              ]
+                0,
+              ],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             0,
             0,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 1,
@@ -1200,53 +1200,53 @@ export const mixedSleet = {
                 "x": [
                   0.36,
                   0.36,
-                  0.36
+                  0.36,
                 ],
                 "y": [
                   1,
                   1,
-                  1
-                ]
+                  1,
+                ],
               },
               "o": {
                 "x": [
                   0.22,
                   0.22,
-                  0.22
+                  0.22,
                 ],
                 "y": [
                   0,
                   0,
-                  0
-                ]
+                  0,
+                ],
               },
               "n": [
                 "0p36_1_0p22_0",
                 "0p36_1_0p22_0",
-                "0p36_1_0p22_0"
+                "0p36_1_0p22_0",
               ],
               "t": 1,
               "s": [
                 100,
                 100,
-                100
+                100,
               ],
               "e": [
                 100,
                 100,
-                100
-              ]
+                100,
+              ],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "ip": -33,
       "op": 849,
       "st": -2,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -1262,53 +1262,53 @@ export const mixedSleet = {
             {
               "i": {
                 "x": [0.518],
-                "y": [0.936]
+                "y": [0.936],
               },
               "o": {
                 "x": [0.294],
-                "y": [0.52]
+                "y": [0.52],
               },
               "n": ["0p518_0p936_0p294_0p52"],
               "t": 1,
               "s": [0],
-              "e": [100]
+              "e": [100],
             },
-            {"t": 37}
+            {"t": 37},
           ],
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 0,
           "k": [
             -250,
             -500,
-            0
+            0,
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             0,
             0,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             100,
             100,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "shapes": [
@@ -1322,21 +1322,21 @@ export const mixedSleet = {
                 "a": 0,
                 "k": [
                   24,
-                  24
+                  24,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "p": {
                 "a": 0,
                 "k": [
                   12,
-                  12
+                  12,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "nm": "Ellipse Path 1",
               "mn": "ADBE Vector Shape - Ellipse",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "fl",
@@ -1346,26 +1346,26 @@ export const mixedSleet = {
                   0.945097982883,
                   0.952941000462,
                   0.956862986088,
-                  1
+                  1,
                 ],
-                "ix": 4
+                "ix": 4,
               },
               "o": {
                 "a": 0,
                 "k": 80,
-                "ix": 5
+                "ix": 5,
               },
               "r": 2,
               "nm": "Fill 1",
               "mn": "ADBE Vector Graphic - Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "gf",
               "o": {
                 "a": 0,
                 "k": 50,
-                "ix": 10
+                "ix": 10,
               },
               "r": 2,
               "g": {
@@ -1390,31 +1390,31 @@ export const mixedSleet = {
                     0.5,
                     0.1,
                     1,
-                    0.2
+                    0.2,
                   ],
-                  "ix": 9
-                }
+                  "ix": 9,
+                },
               },
               "s": {
                 "a": 0,
                 "k": [
                   11.385,
-                  13.566
+                  13.566,
                 ],
-                "ix": 5
+                "ix": 5,
               },
               "e": {
                 "a": 0,
                 "k": [
                   23.401,
-                  24.13
+                  24.13,
                 ],
-                "ix": 6
+                "ix": 6,
               },
               "t": 1,
               "nm": "Gradient Fill 1",
               "mn": "ADBE Vector Graphic - G-Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "tr",
@@ -1422,61 +1422,61 @@ export const mixedSleet = {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "a": {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 1
+                "ix": 1,
               },
               "s": {
                 "a": 0,
                 "k": [
                   3125,
-                  3125
+                  3125,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "r": {
                 "a": 0,
                 "k": 0,
-                "ix": 6
+                "ix": 6,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 7
+                "ix": 7,
               },
               "sk": {
                 "a": 0,
                 "k": 0,
-                "ix": 4
+                "ix": 4,
               },
               "sa": {
                 "a": 0,
                 "k": 0,
-                "ix": 5
+                "ix": 5,
               },
-              "nm": "Transform"
-            }
+              "nm": "Transform",
+            },
           ],
           "nm": "Oval",
           "np": 3,
           "cix": 2,
           "ix": 1,
           "mn": "ADBE Vector Group",
-          "hd": false
-        }
+          "hd": false,
+        },
       ],
       "ip": -58,
       "op": 842,
       "st": -6,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -1491,25 +1491,25 @@ export const mixedSleet = {
             {
               "i": {
                 "x": [0.36],
-                "y": [1]
+                "y": [1],
               },
               "o": {
                 "x": [0.22],
-                "y": [99]
+                "y": [99],
               },
               "n": ["0p36_1_0p22_99"],
               "t": 1,
               "s": [100],
-              "e": [100]
+              "e": [100],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 1,
@@ -1517,57 +1517,57 @@ export const mixedSleet = {
             {
               "i": {
                 "x": 0.36,
-                "y": 1
+                "y": 1,
               },
               "o": {
                 "x": 0.22,
-                "y": 1
+                "y": 1,
               },
               "n": "0p36_1_0p22_1",
               "t": 1,
               "s": [
                 487.75,
                 237,
-                0
+                0,
               ],
               "e": [
                 232.75,
                 237,
-                0
+                0,
               ],
               "to": [
                 -42.5,
                 0,
-                0
+                0,
               ],
               "ti": [
                 42.5,
                 0,
-                0
-              ]
+                0,
+              ],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             -496,
             -494,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             100,
             100,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "shapes": [
@@ -1581,21 +1581,21 @@ export const mixedSleet = {
                 "a": 0,
                 "k": [
                   365.547,
-                  365.547
+                  365.547,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "p": {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "nm": "Ellipse Path 1",
               "mn": "ADBE Vector Shape - Ellipse",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "tr",
@@ -1603,55 +1603,55 @@ export const mixedSleet = {
                 "a": 0,
                 "k": [
                   -499.359,
-                  -499.43
+                  -499.43,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "a": {
                 "a": 0,
                 "k": [
                   -209.552,
-                  -208.357
+                  -208.357,
                 ],
-                "ix": 1
+                "ix": 1,
               },
               "s": {
                 "a": 0,
                 "k": [
                   131.101,
-                  128.967
+                  128.967,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "r": {
                 "a": 0,
                 "k": 0,
-                "ix": 6
+                "ix": 6,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 7
+                "ix": 7,
               },
               "sk": {
                 "a": 0,
                 "k": 0,
-                "ix": 4
+                "ix": 4,
               },
               "sa": {
                 "a": 0,
                 "k": 0,
-                "ix": 5
+                "ix": 5,
               },
-              "nm": "Transform"
-            }
+              "nm": "Transform",
+            },
           ],
           "nm": "Ellipse 1",
           "np": 1,
           "cix": 2,
           "ix": 1,
           "mn": "ADBE Vector Group",
-          "hd": false
+          "hd": false,
         },
         {
           "ty": "st",
@@ -1661,14 +1661,14 @@ export const mixedSleet = {
               0.98431372549,
               0.737254901961,
               0.01568627451,
-              1
+              1,
             ],
-            "ix": 3
+            "ix": 3,
           },
           "o": {
             "a": 0,
             "k": 100,
-            "ix": 4
+            "ix": 4,
           },
           "w": {
             "a": 1,
@@ -1676,33 +1676,33 @@ export const mixedSleet = {
               {
                 "i": {
                   "x": [0.36],
-                  "y": [1]
+                  "y": [1],
                 },
                 "o": {
                   "x": [0.22],
-                  "y": [1]
+                  "y": [1],
                 },
                 "n": ["0p36_1_0p22_1"],
                 "t": 1,
                 "s": [1],
-                "e": [30]
+                "e": [30],
               },
-              {"t": 75}
+              {"t": 75},
             ],
-            "ix": 5
+            "ix": 5,
           },
           "lc": 1,
           "lj": 1,
           "ml": 4,
           "nm": "Stroke 1",
           "mn": "ADBE Vector Graphic - Stroke",
-          "hd": false
-        }
+          "hd": false,
+        },
       ],
       "ip": 1,
       "op": 901,
       "st": -1,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -1716,40 +1716,40 @@ export const mixedSleet = {
         "o": {
           "a": 0,
           "k": 100,
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 0,
           "k": [
             -250,
             -500,
-            0
+            0,
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             0,
             0,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             100,
             100,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "shapes": [
@@ -1763,21 +1763,21 @@ export const mixedSleet = {
                 "a": 0,
                 "k": [
                   24,
-                  24
+                  24,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "p": {
                 "a": 0,
                 "k": [
                   12,
-                  12
+                  12,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "nm": "Ellipse Path 1",
               "mn": "ADBE Vector Shape - Ellipse",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "fl",
@@ -1787,26 +1787,26 @@ export const mixedSleet = {
                   0.945097982883,
                   0.952941000462,
                   0.956862986088,
-                  1
+                  1,
                 ],
-                "ix": 4
+                "ix": 4,
               },
               "o": {
                 "a": 0,
                 "k": 80,
-                "ix": 5
+                "ix": 5,
               },
               "r": 2,
               "nm": "Fill 1",
               "mn": "ADBE Vector Graphic - Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "gf",
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 10
+                "ix": 10,
               },
               "r": 2,
               "g": {
@@ -1831,31 +1831,31 @@ export const mixedSleet = {
                     0.5,
                     0.1,
                     1,
-                    0.2
+                    0.2,
                   ],
-                  "ix": 9
-                }
+                  "ix": 9,
+                },
               },
               "s": {
                 "a": 0,
                 "k": [
                   11.385,
-                  13.566
+                  13.566,
                 ],
-                "ix": 5
+                "ix": 5,
               },
               "e": {
                 "a": 0,
                 "k": [
                   23.401,
-                  24.13
+                  24.13,
                 ],
-                "ix": 6
+                "ix": 6,
               },
               "t": 1,
               "nm": "Gradient Fill 1",
               "mn": "ADBE Vector Graphic - G-Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "tr",
@@ -1863,61 +1863,61 @@ export const mixedSleet = {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "a": {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 1
+                "ix": 1,
               },
               "s": {
                 "a": 0,
                 "k": [
                   3125,
-                  3125
+                  3125,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "r": {
                 "a": 0,
                 "k": 0,
-                "ix": 6
+                "ix": 6,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 7
+                "ix": 7,
               },
               "sk": {
                 "a": 0,
                 "k": 0,
-                "ix": 4
+                "ix": 4,
               },
               "sa": {
                 "a": 0,
                 "k": 0,
-                "ix": 5
+                "ix": 5,
               },
-              "nm": "Transform"
-            }
+              "nm": "Transform",
+            },
           ],
           "nm": "Oval",
           "np": 3,
           "cix": 2,
           "ix": 1,
           "mn": "ADBE Vector Group",
-          "hd": false
-        }
+          "hd": false,
+        },
       ],
       "ip": -58,
       "op": 842,
       "st": -6,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -1934,53 +1934,53 @@ export const mixedSleet = {
             {
               "i": {
                 "x": [0.36],
-                "y": [1]
+                "y": [1],
               },
               "o": {
                 "x": [0.22],
-                "y": [1]
+                "y": [1],
               },
               "n": ["0p36_1_0p22_1"],
               "t": 1,
               "s": [0],
-              "e": [100]
+              "e": [100],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 0,
           "k": [
             -84,
             -62.5,
-            0
+            0,
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             416,
             0,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             100,
             100,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "shapes": [
@@ -1997,52 +1997,52 @@ export const mixedSleet = {
                   "i": [
                     [
                       5.52,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      -5.52
+                      -5.52,
                     ],
                     [
                       0.01,
-                      -0.02
-                    ]
+                      -0.02,
+                    ],
                   ],
                   "o": [
                     [
                       -5.52,
-                      0
+                      0,
                     ],
                     [
                       0.01,
-                      -0.02
+                      -0.02,
                     ],
                     [
                       0,
-                      -5.52
-                    ]
+                      -5.52,
+                    ],
                   ],
                   "v": [
                     [
                       10,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      10
+                      10,
                     ],
                     [
                       20,
-                      10
-                    ]
+                      10,
+                    ],
                   ],
-                  "c": true
+                  "c": true,
                 },
-                "ix": 2
+                "ix": 2,
               },
               "nm": "Path 1",
               "mn": "ADBE Vector Shape - Group",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "fl",
@@ -2052,19 +2052,19 @@ export const mixedSleet = {
                   0.945097982883,
                   0.952941000462,
                   0.956862986088,
-                  1
+                  1,
                 ],
-                "ix": 4
+                "ix": 4,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 5
+                "ix": 5,
               },
               "r": 1,
               "nm": "Fill 1",
               "mn": "ADBE Vector Graphic - Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "tr",
@@ -2072,62 +2072,62 @@ export const mixedSleet = {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "a": {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 1
+                "ix": 1,
               },
               "s": {
                 "a": 0,
                 "k": [
                   3125,
-                  3125
+                  3125,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "r": {
                 "a": 0,
                 "k": 0,
-                "ix": 6
+                "ix": 6,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 7
+                "ix": 7,
               },
               "sk": {
                 "a": 0,
                 "k": 0,
-                "ix": 4
+                "ix": 4,
               },
               "sa": {
                 "a": 0,
                 "k": 0,
-                "ix": 5
+                "ix": 5,
               },
-              "nm": "Transform"
-            }
+              "nm": "Transform",
+            },
           ],
           "nm": "Group 1",
           "np": 2,
           "cix": 2,
           "ix": 1,
           "mn": "ADBE Vector Group",
-          "hd": false
-        }
+          "hd": false,
+        },
       ],
       "ip": -52,
       "op": 848,
       "st": -70,
-      "bm": 0
-    }
+      "bm": 0,
+    },
   ],
-  "markers": []
+  "markers": [],
 };
