@@ -19,12 +19,12 @@ export const mostlyCloudy = {
         "o": {
           "a": 0,
           "k": 0,
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 1,
@@ -32,47 +32,47 @@ export const mostlyCloudy = {
             {
               "i": {
                 "x": 0.36,
-                "y": 1
+                "y": 1,
               },
               "o": {
                 "x": 0.22,
-                "y": 1
+                "y": 1,
               },
               "n": "0p36_1_0p22_1",
               "t": 1,
               "s": [
                 339,
                 667,
-                0
+                0,
               ],
               "e": [
                 667,
                 667,
-                0
+                0,
               ],
               "to": [
                 54.6666679382324,
                 0,
-                0
+                0,
               ],
               "ti": [
                 -54.6666679382324,
                 0,
-                0
-              ]
+                0,
+              ],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             0,
             0,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 1,
@@ -82,53 +82,53 @@ export const mostlyCloudy = {
                 "x": [
                   0.36,
                   0.36,
-                  0.36
+                  0.36,
                 ],
                 "y": [
                   1,
                   1,
-                  1
-                ]
+                  1,
+                ],
               },
               "o": {
                 "x": [
                   0.22,
                   0.22,
-                  0.22
+                  0.22,
                 ],
                 "y": [
                   0,
                   0,
-                  0
-                ]
+                  0,
+                ],
               },
               "n": [
                 "0p36_1_0p22_0",
                 "0p36_1_0p22_0",
-                "0p36_1_0p22_0"
+                "0p36_1_0p22_0",
               ],
               "t": 1,
               "s": [
                 100,
                 100,
-                100
+                100,
               ],
               "e": [
                 100,
                 100,
-                100
-              ]
+                100,
+              ],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "ip": -33,
       "op": 849,
       "st": -2,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -144,53 +144,53 @@ export const mostlyCloudy = {
             {
               "i": {
                 "x": [0.518],
-                "y": [0.936]
+                "y": [0.936],
               },
               "o": {
                 "x": [0.294],
-                "y": [0.52]
+                "y": [0.52],
               },
               "n": ["0p518_0p936_0p294_0p52"],
               "t": 1,
               "s": [0],
-              "e": [100]
+              "e": [100],
             },
-            {"t": 37}
+            {"t": 37},
           ],
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 0,
           "k": [
             -168,
             -166,
-            0
+            0,
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             0,
             0,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             100,
             100,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "shapes": [
@@ -204,21 +204,21 @@ export const mostlyCloudy = {
                 "a": 0,
                 "k": [
                   24,
-                  24
+                  24,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "p": {
                 "a": 0,
                 "k": [
                   12,
-                  12
+                  12,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "nm": "Ellipse Path 1",
               "mn": "ADBE Vector Shape - Ellipse",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "fl",
@@ -228,26 +228,26 @@ export const mostlyCloudy = {
                   0.945097982883,
                   0.952941000462,
                   0.956862986088,
-                  1
+                  1,
                 ],
-                "ix": 4
+                "ix": 4,
               },
               "o": {
                 "a": 0,
                 "k": 80,
-                "ix": 5
+                "ix": 5,
               },
               "r": 2,
               "nm": "Fill 1",
               "mn": "ADBE Vector Graphic - Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "gf",
               "o": {
                 "a": 0,
                 "k": 50,
-                "ix": 10
+                "ix": 10,
               },
               "r": 2,
               "g": {
@@ -272,31 +272,31 @@ export const mostlyCloudy = {
                     0.5,
                     0.1,
                     1,
-                    0.2
+                    0.2,
                   ],
-                  "ix": 9
-                }
+                  "ix": 9,
+                },
               },
               "s": {
                 "a": 0,
                 "k": [
                   11.385,
-                  13.566
+                  13.566,
                 ],
-                "ix": 5
+                "ix": 5,
               },
               "e": {
                 "a": 0,
                 "k": [
                   23.401,
-                  24.13
+                  24.13,
                 ],
-                "ix": 6
+                "ix": 6,
               },
               "t": 1,
               "nm": "Gradient Fill 1",
               "mn": "ADBE Vector Graphic - G-Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "tr",
@@ -304,61 +304,61 @@ export const mostlyCloudy = {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "a": {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 1
+                "ix": 1,
               },
               "s": {
                 "a": 0,
                 "k": [
                   3125,
-                  3125
+                  3125,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "r": {
                 "a": 0,
                 "k": 0,
-                "ix": 6
+                "ix": 6,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 7
+                "ix": 7,
               },
               "sk": {
                 "a": 0,
                 "k": 0,
-                "ix": 4
+                "ix": 4,
               },
               "sa": {
                 "a": 0,
                 "k": 0,
-                "ix": 5
+                "ix": 5,
               },
-              "nm": "Transform"
-            }
+              "nm": "Transform",
+            },
           ],
           "nm": "Oval",
           "np": 3,
           "cix": 2,
           "ix": 1,
           "mn": "ADBE Vector Group",
-          "hd": false
-        }
+          "hd": false,
+        },
       ],
       "ip": -58,
       "op": 842,
       "st": -6,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -372,40 +372,40 @@ export const mostlyCloudy = {
         "o": {
           "a": 0,
           "k": 100,
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 0,
           "k": [
             -168,
             -166,
-            0
+            0,
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             0,
             0,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             100,
             100,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "shapes": [
@@ -419,21 +419,21 @@ export const mostlyCloudy = {
                 "a": 0,
                 "k": [
                   24,
-                  24
+                  24,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "p": {
                 "a": 0,
                 "k": [
                   12,
-                  12
+                  12,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "nm": "Ellipse Path 1",
               "mn": "ADBE Vector Shape - Ellipse",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "fl",
@@ -443,26 +443,26 @@ export const mostlyCloudy = {
                   0.945097982883,
                   0.952941000462,
                   0.956862986088,
-                  1
+                  1,
                 ],
-                "ix": 4
+                "ix": 4,
               },
               "o": {
                 "a": 0,
                 "k": 80,
-                "ix": 5
+                "ix": 5,
               },
               "r": 2,
               "nm": "Fill 1",
               "mn": "ADBE Vector Graphic - Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "gf",
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 10
+                "ix": 10,
               },
               "r": 2,
               "g": {
@@ -487,31 +487,31 @@ export const mostlyCloudy = {
                     0.5,
                     0.1,
                     1,
-                    0.2
+                    0.2,
                   ],
-                  "ix": 9
-                }
+                  "ix": 9,
+                },
               },
               "s": {
                 "a": 0,
                 "k": [
                   11.385,
-                  13.566
+                  13.566,
                 ],
-                "ix": 5
+                "ix": 5,
               },
               "e": {
                 "a": 0,
                 "k": [
                   23.401,
-                  24.13
+                  24.13,
                 ],
-                "ix": 6
+                "ix": 6,
               },
               "t": 1,
               "nm": "Gradient Fill 1",
               "mn": "ADBE Vector Graphic - G-Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "tr",
@@ -519,61 +519,61 @@ export const mostlyCloudy = {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "a": {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 1
+                "ix": 1,
               },
               "s": {
                 "a": 0,
                 "k": [
                   3125,
-                  3125
+                  3125,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "r": {
                 "a": 0,
                 "k": 0,
-                "ix": 6
+                "ix": 6,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 7
+                "ix": 7,
               },
               "sk": {
                 "a": 0,
                 "k": 0,
-                "ix": 4
+                "ix": 4,
               },
               "sa": {
                 "a": 0,
                 "k": 0,
-                "ix": 5
+                "ix": 5,
               },
-              "nm": "Transform"
-            }
+              "nm": "Transform",
+            },
           ],
           "nm": "Oval",
           "np": 3,
           "cix": 2,
           "ix": 1,
           "mn": "ADBE Vector Group",
-          "hd": false
-        }
+          "hd": false,
+        },
       ],
       "ip": -58,
       "op": 842,
       "st": -6,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -590,53 +590,53 @@ export const mostlyCloudy = {
             {
               "i": {
                 "x": [0.36],
-                "y": [1]
+                "y": [1],
               },
               "o": {
                 "x": [0.22],
-                "y": [1]
+                "y": [1],
               },
               "n": ["0p36_1_0p22_1"],
               "t": 1,
               "s": [0],
-              "e": [100]
+              "e": [100],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 0,
           "k": [
             -2,
             271.5,
-            0
+            0,
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             416,
             0,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             100,
             100,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "shapes": [
@@ -653,52 +653,52 @@ export const mostlyCloudy = {
                   "i": [
                     [
                       5.52,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      -5.52
+                      -5.52,
                     ],
                     [
                       0.01,
-                      -0.02
-                    ]
+                      -0.02,
+                    ],
                   ],
                   "o": [
                     [
                       -5.52,
-                      0
+                      0,
                     ],
                     [
                       0.01,
-                      -0.02
+                      -0.02,
                     ],
                     [
                       0,
-                      -5.52
-                    ]
+                      -5.52,
+                    ],
                   ],
                   "v": [
                     [
                       10,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      10
+                      10,
                     ],
                     [
                       20,
-                      10
-                    ]
+                      10,
+                    ],
                   ],
-                  "c": true
+                  "c": true,
                 },
-                "ix": 2
+                "ix": 2,
               },
               "nm": "Path 1",
               "mn": "ADBE Vector Shape - Group",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "fl",
@@ -708,19 +708,19 @@ export const mostlyCloudy = {
                   0.945097982883,
                   0.952941000462,
                   0.956862986088,
-                  1
+                  1,
                 ],
-                "ix": 4
+                "ix": 4,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 5
+                "ix": 5,
               },
               "r": 1,
               "nm": "Fill 1",
               "mn": "ADBE Vector Graphic - Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "tr",
@@ -728,61 +728,61 @@ export const mostlyCloudy = {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "a": {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 1
+                "ix": 1,
               },
               "s": {
                 "a": 0,
                 "k": [
                   3125,
-                  3125
+                  3125,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "r": {
                 "a": 0,
                 "k": 0,
-                "ix": 6
+                "ix": 6,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 7
+                "ix": 7,
               },
               "sk": {
                 "a": 0,
                 "k": 0,
-                "ix": 4
+                "ix": 4,
               },
               "sa": {
                 "a": 0,
                 "k": 0,
-                "ix": 5
+                "ix": 5,
               },
-              "nm": "Transform"
-            }
+              "nm": "Transform",
+            },
           ],
           "nm": "Group 1",
           "np": 2,
           "cix": 2,
           "ix": 1,
           "mn": "ADBE Vector Group",
-          "hd": false
-        }
+          "hd": false,
+        },
       ],
       "ip": -52,
       "op": 848,
       "st": -70,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -794,12 +794,12 @@ export const mostlyCloudy = {
         "o": {
           "a": 0,
           "k": 0,
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 1,
@@ -807,63 +807,63 @@ export const mostlyCloudy = {
             {
               "i": {
                 "x": 0.36,
-                "y": 1
+                "y": 1,
               },
               "o": {
                 "x": 0.22,
-                "y": 1
+                "y": 1,
               },
               "n": "0p36_1_0p22_1",
               "t": 1,
               "s": [
                 171,
                 497,
-                0
+                0,
               ],
               "e": [
                 499,
                 497,
-                0
+                0,
               ],
               "to": [
                 54.6666679382324,
                 0,
-                0
+                0,
               ],
               "ti": [
                 -54.6666679382324,
                 0,
-                0
-              ]
+                0,
+              ],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             0,
             0,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             66,
             66,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "ip": -33,
       "op": 849,
       "st": -2,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -875,12 +875,12 @@ export const mostlyCloudy = {
         "o": {
           "a": 0,
           "k": 0,
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 1,
@@ -888,63 +888,63 @@ export const mostlyCloudy = {
             {
               "i": {
                 "x": 0.36,
-                "y": 1
+                "y": 1,
               },
               "o": {
                 "x": 0.22,
-                "y": 1
+                "y": 1,
               },
               "n": "0p36_1_0p22_1",
               "t": 1,
               "s": [
                 422,
                 750,
-                0
+                0,
               ],
               "e": [
                 750,
                 750,
-                0
+                0,
               ],
               "to": [
                 54.6666679382324,
                 0,
-                0
+                0,
               ],
               "ti": [
                 -54.6666679382324,
                 0,
-                0
-              ]
+                0,
+              ],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             0,
             0,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             65,
             65,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "ip": -17,
       "op": 865,
       "st": 17,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -956,46 +956,46 @@ export const mostlyCloudy = {
         "o": {
           "a": 0,
           "k": 0,
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 0,
           "k": [
             631,
             250,
-            0
+            0,
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             0,
             0,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             76,
             76,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "ip": 0,
       "op": 848,
       "st": 0,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -1011,25 +1011,25 @@ export const mostlyCloudy = {
             {
               "i": {
                 "x": [0.36],
-                "y": [1]
+                "y": [1],
               },
               "o": {
                 "x": [0.22],
-                "y": [1]
+                "y": [1],
               },
               "n": ["0p36_1_0p22_1"],
               "t": 1,
               "s": [1],
-              "e": [100]
+              "e": [100],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 1,
@@ -1037,57 +1037,57 @@ export const mostlyCloudy = {
             {
               "i": {
                 "x": 0.36,
-                "y": 1
+                "y": 1,
               },
               "o": {
                 "x": 0.22,
-                "y": 1
+                "y": 1,
               },
               "n": "0p36_1_0p22_1",
               "t": 1,
               "s": [
                 -169,
                 12,
-                0
+                0,
               ],
               "e": [
                 -494,
                 12,
-                0
+                0,
               ],
               "to": [
                 -54.1666679382324,
                 0,
-                0
+                0,
               ],
               "ti": [
                 54.1666679382324,
                 0,
-                0
-              ]
+                0,
+              ],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             -496,
             -494,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             128,
             128,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "shapes": [
@@ -1101,28 +1101,28 @@ export const mostlyCloudy = {
                 "a": 0,
                 "k": [
                   417.547,
-                  417.547
+                  417.547,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "p": {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "nm": "Ellipse Path 1",
               "mn": "ADBE Vector Shape - Ellipse",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "gf",
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 10
+                "ix": 10,
               },
               "r": 1,
               "g": {
@@ -1141,31 +1141,31 @@ export const mostlyCloudy = {
                     1,
                     0.984,
                     0.737,
-                    0.016
+                    0.016,
                   ],
-                  "ix": 9
-                }
+                  "ix": 9,
+                },
               },
               "s": {
                 "a": 0,
                 "k": [
                   -231.8,
-                  -233.332
+                  -233.332,
                 ],
-                "ix": 5
+                "ix": 5,
               },
               "e": {
                 "a": 0,
                 "k": [
                   14.756,
-                  20.462
+                  20.462,
                 ],
-                "ix": 6
+                "ix": 6,
               },
               "t": 1,
               "nm": "Gradient Fill 1",
               "mn": "ADBE Vector Graphic - G-Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "tr",
@@ -1173,62 +1173,62 @@ export const mostlyCloudy = {
                 "a": 0,
                 "k": [
                   -499.359,
-                  -499.43
+                  -499.43,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "a": {
                 "a": 0,
                 "k": [
                   -209.552,
-                  -208.357
+                  -208.357,
                 ],
-                "ix": 1
+                "ix": 1,
               },
               "s": {
                 "a": 0,
                 "k": [
                   124.101,
-                  122.081
+                  122.081,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "r": {
                 "a": 0,
                 "k": 0,
-                "ix": 6
+                "ix": 6,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 7
+                "ix": 7,
               },
               "sk": {
                 "a": 0,
                 "k": 0,
-                "ix": 4
+                "ix": 4,
               },
               "sa": {
                 "a": 0,
                 "k": 0,
-                "ix": 5
+                "ix": 5,
               },
-              "nm": "Transform"
-            }
+              "nm": "Transform",
+            },
           ],
           "nm": "Ellipse 1",
           "np": 2,
           "cix": 2,
           "ix": 1,
           "mn": "ADBE Vector Group",
-          "hd": false
-        }
+          "hd": false,
+        },
       ],
       "ip": -35,
       "op": 865,
       "st": 17,
-      "bm": 0
-    }
+      "bm": 0,
+    },
   ],
-  "markers": []
+  "markers": [],
 };
