@@ -21,40 +21,40 @@ export const snowShowersDay = {
             "o": {
               "a": 0,
               "k": 100,
-              "ix": 11
+              "ix": 11,
             },
             "r": {
               "a": 0,
               "k": 0,
-              "ix": 10
+              "ix": 10,
             },
             "p": {
               "a": 0,
               "k": [
                 12,
                 12,
-                0
+                0,
               ],
-              "ix": 2
+              "ix": 2,
             },
             "a": {
               "a": 0,
               "k": [
                 8,
                 7,
-                0
+                0,
               ],
-              "ix": 1
+              "ix": 1,
             },
             "s": {
               "a": 0,
               "k": [
                 103,
                 103,
-                100
+                100,
               ],
-              "ix": 6
-            }
+              "ix": 6,
+            },
           },
           "ao": 0,
           "shapes": [
@@ -71,64 +71,64 @@ export const snowShowersDay = {
                       "i": [
                         [
                           -1.039,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          1.039
+                          1.039,
                         ],
                         [
                           1.039,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          -1.039
-                        ]
+                          -1.039,
+                        ],
                       ],
                       "o": [
                         [
                           1.039,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          -1.039
+                          -1.039,
                         ],
                         [
                           -1.039,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          1.039
-                        ]
+                          1.039,
+                        ],
                       ],
                       "v": [
                         [
                           -4.877,
-                          6.757
+                          6.757,
                         ],
                         [
                           -2.996,
-                          4.877
+                          4.877,
                         ],
                         [
                           -4.877,
-                          2.996
+                          2.996,
                         ],
                         [
                           -6.757,
-                          4.877
-                        ]
+                          4.877,
+                        ],
                       ],
-                      "c": true
+                      "c": true,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 17",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ind": 1,
@@ -140,52 +140,52 @@ export const snowShowersDay = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           -1.254,
-                          -1.254
+                          -1.254,
                         ],
                         [
                           -1.254,
-                          1.254
+                          1.254,
                         ],
                         [
                           1.254,
-                          1.254
-                        ]
+                          1.254,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 1",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ind": 2,
@@ -197,52 +197,52 @@ export const snowShowersDay = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           -11.007,
-                          8.499
+                          8.499,
                         ],
                         [
                           -8.499,
-                          8.499
+                          8.499,
                         ],
                         [
                           -8.499,
-                          11.007
-                        ]
+                          11.007,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 2",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ind": 3,
@@ -254,40 +254,40 @@ export const snowShowersDay = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           -8.788,
-                          8.788
+                          8.788,
                         ],
                         [
                           -6.275,
-                          6.275
-                        ]
+                          6.275,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 3",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ind": 4,
@@ -299,40 +299,40 @@ export const snowShowersDay = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           -3.642,
-                          3.642
+                          3.642,
                         ],
                         [
                           -1.129,
-                          1.129
-                        ]
+                          1.129,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 4",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ind": 5,
@@ -344,52 +344,52 @@ export const snowShowersDay = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           2.02,
-                          3.103
+                          3.103,
                         ],
                         [
                           0.247,
-                          4.877
+                          4.877,
                         ],
                         [
                           2.02,
-                          6.65
-                        ]
+                          6.65,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 5",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ind": 6,
@@ -401,52 +401,52 @@ export const snowShowersDay = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           -11.773,
-                          3.103
+                          3.103,
                         ],
                         [
                           -10,
-                          4.877
+                          4.877,
                         ],
                         [
                           -11.773,
-                          6.65
-                        ]
+                          6.65,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 6",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ind": 7,
@@ -458,40 +458,40 @@ export const snowShowersDay = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           -10.312,
-                          4.865
+                          4.865,
                         ],
                         [
                           -6.824,
-                          4.865
-                        ]
+                          4.865,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 7",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ind": 8,
@@ -503,40 +503,40 @@ export const snowShowersDay = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           -3.079,
-                          4.866
+                          4.866,
                         ],
                         [
                           0.409,
-                          4.866
-                        ]
+                          4.866,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 8",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ind": 9,
@@ -548,52 +548,52 @@ export const snowShowersDay = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           1.254,
-                          8.499
+                          8.499,
                         ],
                         [
                           -1.254,
-                          8.499
+                          8.499,
                         ],
                         [
                           -1.254,
-                          11.007
-                        ]
+                          11.007,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 9",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ind": 10,
@@ -605,52 +605,52 @@ export const snowShowersDay = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           -8.499,
-                          -1.254
+                          -1.254,
                         ],
                         [
                           -8.499,
-                          1.254
+                          1.254,
                         ],
                         [
                           -11.007,
-                          1.254
-                        ]
+                          1.254,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 10",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ind": 11,
@@ -662,40 +662,40 @@ export const snowShowersDay = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           -8.63,
-                          1.123
+                          1.123,
                         ],
                         [
                           -6.214,
-                          3.54
-                        ]
+                          3.54,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 11",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ind": 12,
@@ -707,40 +707,40 @@ export const snowShowersDay = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           -3.571,
-                          6.185
+                          6.185,
                         ],
                         [
                           -1.155,
-                          8.601
-                        ]
+                          8.601,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 12",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ind": 13,
@@ -752,52 +752,52 @@ export const snowShowersDay = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           -3.103,
-                          11.773
+                          11.773,
                         ],
                         [
                           -4.877,
-                          10
+                          10,
                         ],
                         [
                           -6.65,
-                          11.773
-                        ]
+                          11.773,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 13",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ind": 14,
@@ -809,52 +809,52 @@ export const snowShowersDay = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           -3.103,
-                          -2.02
+                          -2.02,
                         ],
                         [
                           -4.877,
-                          -0.247
+                          -0.247,
                         ],
                         [
                           -6.65,
-                          -2.02
-                        ]
+                          -2.02,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 14",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ind": 15,
@@ -866,40 +866,40 @@ export const snowShowersDay = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           -4.865,
-                          -0.365
+                          -0.365,
                         ],
                         [
                           -4.865,
-                          2.929
-                        ]
+                          2.929,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 15",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ind": 16,
@@ -911,40 +911,40 @@ export const snowShowersDay = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           -4.866,
-                          6.844
+                          6.844,
                         ],
                         [
                           -4.866,
-                          10.138
-                        ]
+                          10.138,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 16",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ty": "st",
@@ -954,26 +954,26 @@ export const snowShowersDay = {
                       0.258823529412,
                       0.521568627451,
                       0.956862745098,
-                      1
+                      1,
                     ],
-                    "ix": 3
+                    "ix": 3,
                   },
                   "o": {
                     "a": 0,
                     "k": 100,
-                    "ix": 4
+                    "ix": 4,
                   },
                   "w": {
                     "a": 0,
                     "k": 0.909,
-                    "ix": 5
+                    "ix": 5,
                   },
                   "lc": 1,
                   "lj": 1,
                   "ml": 10,
                   "nm": "Stroke 1",
                   "mn": "ADBE Vector Graphic - Stroke",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ty": "tr",
@@ -981,55 +981,55 @@ export const snowShowersDay = {
                     "a": 0,
                     "k": [
                       12.908,
-                      2.155
+                      2.155,
                     ],
-                    "ix": 2
+                    "ix": 2,
                   },
                   "a": {
                     "a": 0,
                     "k": [
                       0,
-                      0
+                      0,
                     ],
-                    "ix": 1
+                    "ix": 1,
                   },
                   "s": {
                     "a": 0,
                     "k": [
                       100,
-                      100
+                      100,
                     ],
-                    "ix": 3
+                    "ix": 3,
                   },
                   "r": {
                     "a": 0,
                     "k": 0,
-                    "ix": 6
+                    "ix": 6,
                   },
                   "o": {
                     "a": 0,
                     "k": 100,
-                    "ix": 7
+                    "ix": 7,
                   },
                   "sk": {
                     "a": 0,
                     "k": 0,
-                    "ix": 4
+                    "ix": 4,
                   },
                   "sa": {
                     "a": 0,
                     "k": 0,
-                    "ix": 5
+                    "ix": 5,
                   },
-                  "nm": "Transform"
-                }
+                  "nm": "Transform",
+                },
               ],
               "nm": "Group 2",
               "np": 18,
               "cix": 2,
               "ix": 1,
               "mn": "ADBE Vector Group",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "gr",
@@ -1044,64 +1044,64 @@ export const snowShowersDay = {
                       "i": [
                         [
                           -1.039,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          1.039
+                          1.039,
                         ],
                         [
                           1.039,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          -1.039
-                        ]
+                          -1.039,
+                        ],
                       ],
                       "o": [
                         [
                           1.039,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          -1.039
+                          -1.039,
                         ],
                         [
                           -1.039,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          1.039
-                        ]
+                          1.039,
+                        ],
                       ],
                       "v": [
                         [
                           8.031,
-                          8.912
+                          8.912,
                         ],
                         [
                           9.912,
-                          7.031
+                          7.031,
                         ],
                         [
                           8.031,
-                          5.15
+                          5.15,
                         ],
                         [
                           6.15,
-                          7.031
-                        ]
+                          7.031,
+                        ],
                       ],
-                      "c": true
+                      "c": true,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 18",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ty": "fl",
@@ -1111,19 +1111,19 @@ export const snowShowersDay = {
                       1,
                       1,
                       1,
-                      1
+                      1,
                     ],
-                    "ix": 4
+                    "ix": 4,
                   },
                   "o": {
                     "a": 0,
                     "k": 100,
-                    "ix": 5
+                    "ix": 5,
                   },
                   "r": 1,
                   "nm": "Fill 1",
                   "mn": "ADBE Vector Graphic - Fill",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ty": "tr",
@@ -1131,64 +1131,64 @@ export const snowShowersDay = {
                     "a": 0,
                     "k": [
                       0,
-                      0
+                      0,
                     ],
-                    "ix": 2
+                    "ix": 2,
                   },
                   "a": {
                     "a": 0,
                     "k": [
                       0,
-                      0
+                      0,
                     ],
-                    "ix": 1
+                    "ix": 1,
                   },
                   "s": {
                     "a": 0,
                     "k": [
                       100,
-                      100
+                      100,
                     ],
-                    "ix": 3
+                    "ix": 3,
                   },
                   "r": {
                     "a": 0,
                     "k": 0,
-                    "ix": 6
+                    "ix": 6,
                   },
                   "o": {
                     "a": 0,
                     "k": 100,
-                    "ix": 7
+                    "ix": 7,
                   },
                   "sk": {
                     "a": 0,
                     "k": 0,
-                    "ix": 4
+                    "ix": 4,
                   },
                   "sa": {
                     "a": 0,
                     "k": 0,
-                    "ix": 5
+                    "ix": 5,
                   },
-                  "nm": "Transform"
-                }
+                  "nm": "Transform",
+                },
               ],
               "nm": "Group 1",
               "np": 2,
               "cix": 2,
               "ix": 2,
               "mn": "ADBE Vector Group",
-              "hd": false
-            }
+              "hd": false,
+            },
           ],
           "ip": 0,
           "op": 848,
           "st": 0,
-          "bm": 0
-        }
-      ]
-    }
+          "bm": 0,
+        },
+      ],
+    },
   ],
   "layers": [
     {
@@ -1205,20 +1205,20 @@ export const snowShowersDay = {
             {
               "i": {
                 "x": [0.833],
-                "y": [0.833]
+                "y": [0.833],
               },
               "o": {
                 "x": [0.167],
-                "y": [0.167]
+                "y": [0.167],
               },
               "n": ["0p833_0p833_0p167_0p167"],
               "t": 29,
               "s": [0],
-              "e": [100]
+              "e": [100],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 1,
@@ -1226,20 +1226,20 @@ export const snowShowersDay = {
             {
               "i": {
                 "x": [0.36],
-                "y": [1]
+                "y": [1],
               },
               "o": {
                 "x": [0.22],
-                "y": [1]
+                "y": [1],
               },
               "n": ["0p36_1_0p22_1"],
               "t": 29,
               "s": [-45],
-              "e": [0]
+              "e": [0],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 1,
@@ -1247,57 +1247,57 @@ export const snowShowersDay = {
             {
               "i": {
                 "x": 0.36,
-                "y": 1
+                "y": 1,
               },
               "o": {
                 "x": 0.22,
-                "y": 1
+                "y": 1,
               },
               "n": "0p36_1_0p22_1",
               "t": 29,
               "s": [
                 1005,
                 713,
-                0
+                0,
               ],
               "e": [
                 1005,
                 981,
-                0
+                0,
               ],
               "to": [
                 0,
                 44.6666679382324,
-                0
+                0,
               ],
               "ti": [
                 0,
                 -44.6666679382324,
-                0
-              ]
+                0,
+              ],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             12,
             12,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             3125,
             3125,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "w": 24,
@@ -1305,7 +1305,7 @@ export const snowShowersDay = {
       "ip": 29,
       "op": 877,
       "st": 29,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -1317,12 +1317,12 @@ export const snowShowersDay = {
         "o": {
           "a": 0,
           "k": 0,
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 1,
@@ -1330,47 +1330,47 @@ export const snowShowersDay = {
             {
               "i": {
                 "x": 0.36,
-                "y": 1
+                "y": 1,
               },
               "o": {
                 "x": 0.22,
-                "y": 1
+                "y": 1,
               },
               "n": "0p36_1_0p22_1",
               "t": 1,
               "s": [
                 422,
                 750,
-                0
+                0,
               ],
               "e": [
                 750,
                 750,
-                0
+                0,
               ],
               "to": [
                 54.6666679382324,
                 0,
-                0
+                0,
               ],
               "ti": [
                 -54.6666679382324,
                 0,
-                0
-              ]
+                0,
+              ],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             0,
             0,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 1,
@@ -1380,53 +1380,53 @@ export const snowShowersDay = {
                 "x": [
                   0.36,
                   0.36,
-                  0.36
+                  0.36,
                 ],
                 "y": [
                   1,
                   1,
-                  1
-                ]
+                  1,
+                ],
               },
               "o": {
                 "x": [
                   0.22,
                   0.22,
-                  0.22
+                  0.22,
                 ],
                 "y": [
                   0,
                   0,
-                  0
-                ]
+                  0,
+                ],
               },
               "n": [
                 "0p36_1_0p22_0",
                 "0p36_1_0p22_0",
-                "0p36_1_0p22_0"
+                "0p36_1_0p22_0",
               ],
               "t": 1,
               "s": [
                 100,
                 100,
-                100
+                100,
               ],
               "e": [
                 100,
                 100,
-                100
-              ]
+                100,
+              ],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "ip": -33,
       "op": 849,
       "st": -2,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -1442,53 +1442,53 @@ export const snowShowersDay = {
             {
               "i": {
                 "x": [0.518],
-                "y": [0.936]
+                "y": [0.936],
               },
               "o": {
                 "x": [0.294],
-                "y": [0.52]
+                "y": [0.52],
               },
               "n": ["0p518_0p936_0p294_0p52"],
               "t": 1,
               "s": [0],
-              "e": [100]
+              "e": [100],
             },
-            {"t": 37}
+            {"t": 37},
           ],
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 0,
           "k": [
             -250,
             -500,
-            0
+            0,
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             0,
             0,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             100,
             100,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "shapes": [
@@ -1502,21 +1502,21 @@ export const snowShowersDay = {
                 "a": 0,
                 "k": [
                   24,
-                  24
+                  24,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "p": {
                 "a": 0,
                 "k": [
                   12,
-                  12
+                  12,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "nm": "Ellipse Path 1",
               "mn": "ADBE Vector Shape - Ellipse",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "fl",
@@ -1526,26 +1526,26 @@ export const snowShowersDay = {
                   0.945097982883,
                   0.952941000462,
                   0.956862986088,
-                  1
+                  1,
                 ],
-                "ix": 4
+                "ix": 4,
               },
               "o": {
                 "a": 0,
                 "k": 80,
-                "ix": 5
+                "ix": 5,
               },
               "r": 2,
               "nm": "Fill 1",
               "mn": "ADBE Vector Graphic - Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "gf",
               "o": {
                 "a": 0,
                 "k": 50,
-                "ix": 10
+                "ix": 10,
               },
               "r": 2,
               "g": {
@@ -1570,31 +1570,31 @@ export const snowShowersDay = {
                     0.5,
                     0.1,
                     1,
-                    0.2
+                    0.2,
                   ],
-                  "ix": 9
-                }
+                  "ix": 9,
+                },
               },
               "s": {
                 "a": 0,
                 "k": [
                   11.385,
-                  13.566
+                  13.566,
                 ],
-                "ix": 5
+                "ix": 5,
               },
               "e": {
                 "a": 0,
                 "k": [
                   23.401,
-                  24.13
+                  24.13,
                 ],
-                "ix": 6
+                "ix": 6,
               },
               "t": 1,
               "nm": "Gradient Fill 1",
               "mn": "ADBE Vector Graphic - G-Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "tr",
@@ -1602,61 +1602,61 @@ export const snowShowersDay = {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "a": {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 1
+                "ix": 1,
               },
               "s": {
                 "a": 0,
                 "k": [
                   3125,
-                  3125
+                  3125,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "r": {
                 "a": 0,
                 "k": 0,
-                "ix": 6
+                "ix": 6,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 7
+                "ix": 7,
               },
               "sk": {
                 "a": 0,
                 "k": 0,
-                "ix": 4
+                "ix": 4,
               },
               "sa": {
                 "a": 0,
                 "k": 0,
-                "ix": 5
+                "ix": 5,
               },
-              "nm": "Transform"
-            }
+              "nm": "Transform",
+            },
           ],
           "nm": "Oval",
           "np": 3,
           "cix": 2,
           "ix": 1,
           "mn": "ADBE Vector Group",
-          "hd": false
-        }
+          "hd": false,
+        },
       ],
       "ip": -58,
       "op": 842,
       "st": -6,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -1671,25 +1671,25 @@ export const snowShowersDay = {
             {
               "i": {
                 "x": [0.36],
-                "y": [1]
+                "y": [1],
               },
               "o": {
                 "x": [0.22],
-                "y": [99]
+                "y": [99],
               },
               "n": ["0p36_1_0p22_99"],
               "t": 1,
               "s": [100],
-              "e": [100]
+              "e": [100],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 1,
@@ -1697,57 +1697,57 @@ export const snowShowersDay = {
             {
               "i": {
                 "x": 0.36,
-                "y": 1
+                "y": 1,
               },
               "o": {
                 "x": 0.22,
-                "y": 1
+                "y": 1,
               },
               "n": "0p36_1_0p22_1",
               "t": 1,
               "s": [
                 487.75,
                 237,
-                0
+                0,
               ],
               "e": [
                 232.75,
                 237,
-                0
+                0,
               ],
               "to": [
                 -42.5,
                 0,
-                0
+                0,
               ],
               "ti": [
                 42.5,
                 0,
-                0
-              ]
+                0,
+              ],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             -496,
             -494,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             100,
             100,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "shapes": [
@@ -1761,21 +1761,21 @@ export const snowShowersDay = {
                 "a": 0,
                 "k": [
                   365.547,
-                  365.547
+                  365.547,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "p": {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "nm": "Ellipse Path 1",
               "mn": "ADBE Vector Shape - Ellipse",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "tr",
@@ -1783,55 +1783,55 @@ export const snowShowersDay = {
                 "a": 0,
                 "k": [
                   -499.359,
-                  -499.43
+                  -499.43,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "a": {
                 "a": 0,
                 "k": [
                   -209.552,
-                  -208.357
+                  -208.357,
                 ],
-                "ix": 1
+                "ix": 1,
               },
               "s": {
                 "a": 0,
                 "k": [
                   131.101,
-                  128.967
+                  128.967,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "r": {
                 "a": 0,
                 "k": 0,
-                "ix": 6
+                "ix": 6,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 7
+                "ix": 7,
               },
               "sk": {
                 "a": 0,
                 "k": 0,
-                "ix": 4
+                "ix": 4,
               },
               "sa": {
                 "a": 0,
                 "k": 0,
-                "ix": 5
+                "ix": 5,
               },
-              "nm": "Transform"
-            }
+              "nm": "Transform",
+            },
           ],
           "nm": "Ellipse 1",
           "np": 1,
           "cix": 2,
           "ix": 1,
           "mn": "ADBE Vector Group",
-          "hd": false
+          "hd": false,
         },
         {
           "ty": "st",
@@ -1841,14 +1841,14 @@ export const snowShowersDay = {
               0.98431372549,
               0.737254901961,
               0.01568627451,
-              1
+              1,
             ],
-            "ix": 3
+            "ix": 3,
           },
           "o": {
             "a": 0,
             "k": 100,
-            "ix": 4
+            "ix": 4,
           },
           "w": {
             "a": 1,
@@ -1856,33 +1856,33 @@ export const snowShowersDay = {
               {
                 "i": {
                   "x": [0.36],
-                  "y": [1]
+                  "y": [1],
                 },
                 "o": {
                   "x": [0.22],
-                  "y": [1]
+                  "y": [1],
                 },
                 "n": ["0p36_1_0p22_1"],
                 "t": 1,
                 "s": [1],
-                "e": [30]
+                "e": [30],
               },
-              {"t": 75}
+              {"t": 75},
             ],
-            "ix": 5
+            "ix": 5,
           },
           "lc": 1,
           "lj": 1,
           "ml": 4,
           "nm": "Stroke 1",
           "mn": "ADBE Vector Graphic - Stroke",
-          "hd": false
-        }
+          "hd": false,
+        },
       ],
       "ip": 1,
       "op": 901,
       "st": -1,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -1896,40 +1896,40 @@ export const snowShowersDay = {
         "o": {
           "a": 0,
           "k": 100,
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 0,
           "k": [
             -250,
             -500,
-            0
+            0,
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             0,
             0,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             100,
             100,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "shapes": [
@@ -1943,21 +1943,21 @@ export const snowShowersDay = {
                 "a": 0,
                 "k": [
                   24,
-                  24
+                  24,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "p": {
                 "a": 0,
                 "k": [
                   12,
-                  12
+                  12,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "nm": "Ellipse Path 1",
               "mn": "ADBE Vector Shape - Ellipse",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "fl",
@@ -1967,26 +1967,26 @@ export const snowShowersDay = {
                   0.945097982883,
                   0.952941000462,
                   0.956862986088,
-                  1
+                  1,
                 ],
-                "ix": 4
+                "ix": 4,
               },
               "o": {
                 "a": 0,
                 "k": 80,
-                "ix": 5
+                "ix": 5,
               },
               "r": 2,
               "nm": "Fill 1",
               "mn": "ADBE Vector Graphic - Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "gf",
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 10
+                "ix": 10,
               },
               "r": 2,
               "g": {
@@ -2011,31 +2011,31 @@ export const snowShowersDay = {
                     0.5,
                     0.1,
                     1,
-                    0.2
+                    0.2,
                   ],
-                  "ix": 9
-                }
+                  "ix": 9,
+                },
               },
               "s": {
                 "a": 0,
                 "k": [
                   11.385,
-                  13.566
+                  13.566,
                 ],
-                "ix": 5
+                "ix": 5,
               },
               "e": {
                 "a": 0,
                 "k": [
                   23.401,
-                  24.13
+                  24.13,
                 ],
-                "ix": 6
+                "ix": 6,
               },
               "t": 1,
               "nm": "Gradient Fill 1",
               "mn": "ADBE Vector Graphic - G-Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "tr",
@@ -2043,61 +2043,61 @@ export const snowShowersDay = {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "a": {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 1
+                "ix": 1,
               },
               "s": {
                 "a": 0,
                 "k": [
                   3125,
-                  3125
+                  3125,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "r": {
                 "a": 0,
                 "k": 0,
-                "ix": 6
+                "ix": 6,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 7
+                "ix": 7,
               },
               "sk": {
                 "a": 0,
                 "k": 0,
-                "ix": 4
+                "ix": 4,
               },
               "sa": {
                 "a": 0,
                 "k": 0,
-                "ix": 5
+                "ix": 5,
               },
-              "nm": "Transform"
-            }
+              "nm": "Transform",
+            },
           ],
           "nm": "Oval",
           "np": 3,
           "cix": 2,
           "ix": 1,
           "mn": "ADBE Vector Group",
-          "hd": false
-        }
+          "hd": false,
+        },
       ],
       "ip": -58,
       "op": 842,
       "st": -6,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -2114,53 +2114,53 @@ export const snowShowersDay = {
             {
               "i": {
                 "x": [0.36],
-                "y": [1]
+                "y": [1],
               },
               "o": {
                 "x": [0.22],
-                "y": [1]
+                "y": [1],
               },
               "n": ["0p36_1_0p22_1"],
               "t": 1,
               "s": [0],
-              "e": [100]
+              "e": [100],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 0,
           "k": [
             -84,
             -62.5,
-            0
+            0,
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             416,
             0,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             100,
             100,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "shapes": [
@@ -2177,52 +2177,52 @@ export const snowShowersDay = {
                   "i": [
                     [
                       5.52,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      -5.52
+                      -5.52,
                     ],
                     [
                       0.01,
-                      -0.02
-                    ]
+                      -0.02,
+                    ],
                   ],
                   "o": [
                     [
                       -5.52,
-                      0
+                      0,
                     ],
                     [
                       0.01,
-                      -0.02
+                      -0.02,
                     ],
                     [
                       0,
-                      -5.52
-                    ]
+                      -5.52,
+                    ],
                   ],
                   "v": [
                     [
                       10,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      10
+                      10,
                     ],
                     [
                       20,
-                      10
-                    ]
+                      10,
+                    ],
                   ],
-                  "c": true
+                  "c": true,
                 },
-                "ix": 2
+                "ix": 2,
               },
               "nm": "Path 1",
               "mn": "ADBE Vector Shape - Group",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "fl",
@@ -2232,19 +2232,19 @@ export const snowShowersDay = {
                   0.945097982883,
                   0.952941000462,
                   0.956862986088,
-                  1
+                  1,
                 ],
-                "ix": 4
+                "ix": 4,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 5
+                "ix": 5,
               },
               "r": 1,
               "nm": "Fill 1",
               "mn": "ADBE Vector Graphic - Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "tr",
@@ -2252,62 +2252,62 @@ export const snowShowersDay = {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "a": {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 1
+                "ix": 1,
               },
               "s": {
                 "a": 0,
                 "k": [
                   3125,
-                  3125
+                  3125,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "r": {
                 "a": 0,
                 "k": 0,
-                "ix": 6
+                "ix": 6,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 7
+                "ix": 7,
               },
               "sk": {
                 "a": 0,
                 "k": 0,
-                "ix": 4
+                "ix": 4,
               },
               "sa": {
                 "a": 0,
                 "k": 0,
-                "ix": 5
+                "ix": 5,
               },
-              "nm": "Transform"
-            }
+              "nm": "Transform",
+            },
           ],
           "nm": "Group 1",
           "np": 2,
           "cix": 2,
           "ix": 1,
           "mn": "ADBE Vector Group",
-          "hd": false
-        }
+          "hd": false,
+        },
       ],
       "ip": -52,
       "op": 848,
       "st": -70,
-      "bm": 0
-    }
+      "bm": 0,
+    },
   ],
-  "markers": []
+  "markers": [],
 };
