@@ -21,40 +21,40 @@ export const winteryMixDay = {
             "o": {
               "a": 0,
               "k": 100,
-              "ix": 11
+              "ix": 11,
             },
             "r": {
               "a": 0,
               "k": -45,
-              "ix": 10
+              "ix": 10,
             },
             "p": {
               "a": 0,
               "k": [
                 -15,
                 122.5,
-                0
+                0,
               ],
-              "ix": 2
+              "ix": 2,
             },
             "a": {
               "a": 0,
               "k": [
                 0,
                 0,
-                0
+                0,
               ],
-              "ix": 1
+              "ix": 1,
             },
             "s": {
               "a": 0,
               "k": [
                 1062.5,
                 1062.5,
-                100
+                100,
               ],
-              "ix": 6
-            }
+              "ix": 6,
+            },
           },
           "ao": 0,
           "shapes": [
@@ -71,21 +71,21 @@ export const winteryMixDay = {
                         "a": 0,
                         "k": [
                           16,
-                          16
+                          16,
                         ],
-                        "ix": 2
+                        "ix": 2,
                       },
                       "p": {
                         "a": 0,
                         "k": [
                           8,
-                          8
+                          8,
                         ],
-                        "ix": 3
+                        "ix": 3,
                       },
                       "nm": "Ellipse Path 1",
                       "mn": "ADBE Vector Shape - Ellipse",
-                      "hd": false
+                      "hd": false,
                     },
                     {
                       "ty": "tr",
@@ -93,55 +93,55 @@ export const winteryMixDay = {
                         "a": 0,
                         "k": [
                           8,
-                          8
+                          8,
                         ],
-                        "ix": 2
+                        "ix": 2,
                       },
                       "a": {
                         "a": 0,
                         "k": [
                           8,
-                          8
+                          8,
                         ],
-                        "ix": 1
+                        "ix": 1,
                       },
                       "s": {
                         "a": 0,
                         "k": [
                           100,
-                          100
+                          100,
                         ],
-                        "ix": 3
+                        "ix": 3,
                       },
                       "r": {
                         "a": 0,
                         "k": 0,
-                        "ix": 6
+                        "ix": 6,
                       },
                       "o": {
                         "a": 0,
                         "k": 100,
-                        "ix": 7
+                        "ix": 7,
                       },
                       "sk": {
                         "a": 0,
                         "k": 0,
-                        "ix": 4
+                        "ix": 4,
                       },
                       "sa": {
                         "a": 0,
                         "k": 0,
-                        "ix": 5
+                        "ix": 5,
                       },
-                      "nm": "Transform"
-                    }
+                      "nm": "Transform",
+                    },
                   ],
                   "nm": "Oval",
                   "np": 1,
                   "cix": 2,
                   "ix": 1,
                   "mn": "ADBE Vector Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ty": "gr",
@@ -153,26 +153,26 @@ export const winteryMixDay = {
                         "a": 0,
                         "k": [
                           8,
-                          8
+                          8,
                         ],
-                        "ix": 2
+                        "ix": 2,
                       },
                       "p": {
                         "a": 0,
                         "k": [
                           12,
-                          4
+                          4,
                         ],
-                        "ix": 3
+                        "ix": 3,
                       },
                       "r": {
                         "a": 0,
                         "k": 0,
-                        "ix": 4
+                        "ix": 4,
                       },
                       "nm": "Rectangle Path 1",
                       "mn": "ADBE Vector Shape - Rect",
-                      "hd": false
+                      "hd": false,
                     },
                     {
                       "ty": "tr",
@@ -180,55 +180,55 @@ export const winteryMixDay = {
                         "a": 0,
                         "k": [
                           12,
-                          4
+                          4,
                         ],
-                        "ix": 2
+                        "ix": 2,
                       },
                       "a": {
                         "a": 0,
                         "k": [
                           12,
-                          4
+                          4,
                         ],
-                        "ix": 1
+                        "ix": 1,
                       },
                       "s": {
                         "a": 0,
                         "k": [
                           100,
-                          100
+                          100,
                         ],
-                        "ix": 3
+                        "ix": 3,
                       },
                       "r": {
                         "a": 0,
                         "k": 0,
-                        "ix": 6
+                        "ix": 6,
                       },
                       "o": {
                         "a": 0,
                         "k": 100,
-                        "ix": 7
+                        "ix": 7,
                       },
                       "sk": {
                         "a": 0,
                         "k": 0,
-                        "ix": 4
+                        "ix": 4,
                       },
                       "sa": {
                         "a": 0,
                         "k": 0,
-                        "ix": 5
+                        "ix": 5,
                       },
-                      "nm": "Transform"
-                    }
+                      "nm": "Transform",
+                    },
                   ],
                   "nm": "Rectangle 2",
                   "np": 1,
                   "cix": 2,
                   "ix": 2,
                   "mn": "ADBE Vector Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ty": "fl",
@@ -238,19 +238,19 @@ export const winteryMixDay = {
                       0.258823990822,
                       0.521569013596,
                       0.956862986088,
-                      1
+                      1,
                     ],
-                    "ix": 4
+                    "ix": 4,
                   },
                   "o": {
                     "a": 0,
                     "k": 100,
-                    "ix": 5
+                    "ix": 5,
                   },
                   "r": 1,
                   "nm": "Fill 1",
                   "mn": "ADBE Vector Graphic - Fill",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ty": "tr",
@@ -258,63 +258,63 @@ export const winteryMixDay = {
                     "a": 0,
                     "k": [
                       0,
-                      0
+                      0,
                     ],
-                    "ix": 2
+                    "ix": 2,
                   },
                   "a": {
                     "a": 0,
                     "k": [
                       0,
-                      0
+                      0,
                     ],
-                    "ix": 1
+                    "ix": 1,
                   },
                   "s": {
                     "a": 0,
                     "k": [
                       100,
-                      100
+                      100,
                     ],
-                    "ix": 3
+                    "ix": 3,
                   },
                   "r": {
                     "a": 0,
                     "k": 0,
-                    "ix": 6
+                    "ix": 6,
                   },
                   "o": {
                     "a": 0,
                     "k": 100,
-                    "ix": 7
+                    "ix": 7,
                   },
                   "sk": {
                     "a": 0,
                     "k": 0,
-                    "ix": 4
+                    "ix": 4,
                   },
                   "sa": {
                     "a": 0,
                     "k": 0,
-                    "ix": 5
+                    "ix": 5,
                   },
-                  "nm": "Transform"
-                }
+                  "nm": "Transform",
+                },
               ],
               "nm": "Combined Shape Copy",
               "np": 3,
               "cix": 2,
               "ix": 1,
               "mn": "ADBE Vector Group",
-              "hd": false
-            }
+              "hd": false,
+            },
           ],
           "ip": 0,
           "op": 848,
           "st": 0,
-          "bm": 0
-        }
-      ]
+          "bm": 0,
+        },
+      ],
     },
     {
       "id": "comp_1089",
@@ -329,40 +329,40 @@ export const winteryMixDay = {
             "o": {
               "a": 0,
               "k": 100,
-              "ix": 11
+              "ix": 11,
             },
             "r": {
               "a": 0,
               "k": -45,
-              "ix": 10
+              "ix": 10,
             },
             "p": {
               "a": 0,
               "k": [
                 104.5,
                 105.5,
-                0
+                0,
               ],
-              "ix": 2
+              "ix": 2,
             },
             "a": {
               "a": 0,
               "k": [
                 7.953,
                 7.953,
-                0
+                0,
               ],
-              "ix": 1
+              "ix": 1,
             },
             "s": {
               "a": 0,
               "k": [
                 1031.5,
                 1031.5,
-                100
+                100,
               ],
-              "ix": 6
-            }
+              "ix": 6,
+            },
           },
           "ao": 0,
           "shapes": [
@@ -379,21 +379,21 @@ export const winteryMixDay = {
                         "a": 0,
                         "k": [
                           16,
-                          16
+                          16,
                         ],
-                        "ix": 2
+                        "ix": 2,
                       },
                       "p": {
                         "a": 0,
                         "k": [
                           8,
-                          8
+                          8,
                         ],
-                        "ix": 3
+                        "ix": 3,
                       },
                       "nm": "Ellipse Path 1",
                       "mn": "ADBE Vector Shape - Ellipse",
-                      "hd": false
+                      "hd": false,
                     },
                     {
                       "ty": "tr",
@@ -401,55 +401,55 @@ export const winteryMixDay = {
                         "a": 0,
                         "k": [
                           8,
-                          8
+                          8,
                         ],
-                        "ix": 2
+                        "ix": 2,
                       },
                       "a": {
                         "a": 0,
                         "k": [
                           8,
-                          8
+                          8,
                         ],
-                        "ix": 1
+                        "ix": 1,
                       },
                       "s": {
                         "a": 0,
                         "k": [
                           100,
-                          100
+                          100,
                         ],
-                        "ix": 3
+                        "ix": 3,
                       },
                       "r": {
                         "a": 0,
                         "k": 0,
-                        "ix": 6
+                        "ix": 6,
                       },
                       "o": {
                         "a": 0,
                         "k": 100,
-                        "ix": 7
+                        "ix": 7,
                       },
                       "sk": {
                         "a": 0,
                         "k": 0,
-                        "ix": 4
+                        "ix": 4,
                       },
                       "sa": {
                         "a": 0,
                         "k": 0,
-                        "ix": 5
+                        "ix": 5,
                       },
-                      "nm": "Transform"
-                    }
+                      "nm": "Transform",
+                    },
                   ],
                   "nm": "Oval",
                   "np": 1,
                   "cix": 2,
                   "ix": 1,
                   "mn": "ADBE Vector Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ty": "tr",
@@ -457,55 +457,55 @@ export const winteryMixDay = {
                     "a": 0,
                     "k": [
                       0,
-                      0
+                      0,
                     ],
-                    "ix": 2
+                    "ix": 2,
                   },
                   "a": {
                     "a": 0,
                     "k": [
                       0,
-                      0
+                      0,
                     ],
-                    "ix": 1
+                    "ix": 1,
                   },
                   "s": {
                     "a": 0,
                     "k": [
                       100,
-                      100
+                      100,
                     ],
-                    "ix": 3
+                    "ix": 3,
                   },
                   "r": {
                     "a": 0,
                     "k": 0,
-                    "ix": 6
+                    "ix": 6,
                   },
                   "o": {
                     "a": 0,
                     "k": 100,
-                    "ix": 7
+                    "ix": 7,
                   },
                   "sk": {
                     "a": 0,
                     "k": 0,
-                    "ix": 4
+                    "ix": 4,
                   },
                   "sa": {
                     "a": 0,
                     "k": 0,
-                    "ix": 5
+                    "ix": 5,
                   },
-                  "nm": "Transform"
-                }
+                  "nm": "Transform",
+                },
               ],
               "nm": "Combined Shape Copy",
               "np": 2,
               "cix": 2,
               "ix": 1,
               "mn": "ADBE Vector Group",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "st",
@@ -515,34 +515,34 @@ export const winteryMixDay = {
                   0.258823529412,
                   0.521568627451,
                   0.956862745098,
-                  1
+                  1,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 4
+                "ix": 4,
               },
               "w": {
                 "a": 0,
                 "k": 3,
-                "ix": 5
+                "ix": 5,
               },
               "lc": 1,
               "lj": 1,
               "ml": 4,
               "nm": "Stroke 1",
               "mn": "ADBE Vector Graphic - Stroke",
-              "hd": false
-            }
+              "hd": false,
+            },
           ],
           "ip": 0,
           "op": 848,
           "st": 0,
-          "bm": 0
-        }
-      ]
+          "bm": 0,
+        },
+      ],
     },
     {
       "id": "comp_1090",
@@ -557,40 +557,40 @@ export const winteryMixDay = {
             "o": {
               "a": 0,
               "k": 100,
-              "ix": 11
+              "ix": 11,
             },
             "r": {
               "a": 0,
               "k": 0,
-              "ix": 10
+              "ix": 10,
             },
             "p": {
               "a": 0,
               "k": [
                 12,
                 12,
-                0
+                0,
               ],
-              "ix": 2
+              "ix": 2,
             },
             "a": {
               "a": 0,
               "k": [
                 8,
                 7,
-                0
+                0,
               ],
-              "ix": 1
+              "ix": 1,
             },
             "s": {
               "a": 0,
               "k": [
                 103,
                 103,
-                100
+                100,
               ],
-              "ix": 6
-            }
+              "ix": 6,
+            },
           },
           "ao": 0,
           "shapes": [
@@ -607,64 +607,64 @@ export const winteryMixDay = {
                       "i": [
                         [
                           -1.039,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          1.039
+                          1.039,
                         ],
                         [
                           1.039,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          -1.039
-                        ]
+                          -1.039,
+                        ],
                       ],
                       "o": [
                         [
                           1.039,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          -1.039
+                          -1.039,
                         ],
                         [
                           -1.039,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          1.039
-                        ]
+                          1.039,
+                        ],
                       ],
                       "v": [
                         [
                           -4.877,
-                          6.757
+                          6.757,
                         ],
                         [
                           -2.996,
-                          4.877
+                          4.877,
                         ],
                         [
                           -4.877,
-                          2.996
+                          2.996,
                         ],
                         [
                           -6.757,
-                          4.877
-                        ]
+                          4.877,
+                        ],
                       ],
-                      "c": true
+                      "c": true,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 17",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ind": 1,
@@ -676,52 +676,52 @@ export const winteryMixDay = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           -1.254,
-                          -1.254
+                          -1.254,
                         ],
                         [
                           -1.254,
-                          1.254
+                          1.254,
                         ],
                         [
                           1.254,
-                          1.254
-                        ]
+                          1.254,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 1",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ind": 2,
@@ -733,52 +733,52 @@ export const winteryMixDay = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           -11.007,
-                          8.499
+                          8.499,
                         ],
                         [
                           -8.499,
-                          8.499
+                          8.499,
                         ],
                         [
                           -8.499,
-                          11.007
-                        ]
+                          11.007,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 2",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ind": 3,
@@ -790,40 +790,40 @@ export const winteryMixDay = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           -8.788,
-                          8.788
+                          8.788,
                         ],
                         [
                           -6.275,
-                          6.275
-                        ]
+                          6.275,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 3",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ind": 4,
@@ -835,40 +835,40 @@ export const winteryMixDay = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           -3.642,
-                          3.642
+                          3.642,
                         ],
                         [
                           -1.129,
-                          1.129
-                        ]
+                          1.129,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 4",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ind": 5,
@@ -880,52 +880,52 @@ export const winteryMixDay = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           2.02,
-                          3.103
+                          3.103,
                         ],
                         [
                           0.247,
-                          4.877
+                          4.877,
                         ],
                         [
                           2.02,
-                          6.65
-                        ]
+                          6.65,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 5",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ind": 6,
@@ -937,52 +937,52 @@ export const winteryMixDay = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           -11.773,
-                          3.103
+                          3.103,
                         ],
                         [
                           -10,
-                          4.877
+                          4.877,
                         ],
                         [
                           -11.773,
-                          6.65
-                        ]
+                          6.65,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 6",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ind": 7,
@@ -994,40 +994,40 @@ export const winteryMixDay = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           -10.312,
-                          4.865
+                          4.865,
                         ],
                         [
                           -6.824,
-                          4.865
-                        ]
+                          4.865,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 7",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ind": 8,
@@ -1039,40 +1039,40 @@ export const winteryMixDay = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           -3.079,
-                          4.866
+                          4.866,
                         ],
                         [
                           0.409,
-                          4.866
-                        ]
+                          4.866,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 8",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ind": 9,
@@ -1084,52 +1084,52 @@ export const winteryMixDay = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           1.254,
-                          8.499
+                          8.499,
                         ],
                         [
                           -1.254,
-                          8.499
+                          8.499,
                         ],
                         [
                           -1.254,
-                          11.007
-                        ]
+                          11.007,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 9",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ind": 10,
@@ -1141,52 +1141,52 @@ export const winteryMixDay = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           -8.499,
-                          -1.254
+                          -1.254,
                         ],
                         [
                           -8.499,
-                          1.254
+                          1.254,
                         ],
                         [
                           -11.007,
-                          1.254
-                        ]
+                          1.254,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 10",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ind": 11,
@@ -1198,40 +1198,40 @@ export const winteryMixDay = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           -8.63,
-                          1.123
+                          1.123,
                         ],
                         [
                           -6.214,
-                          3.54
-                        ]
+                          3.54,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 11",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ind": 12,
@@ -1243,40 +1243,40 @@ export const winteryMixDay = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           -3.571,
-                          6.185
+                          6.185,
                         ],
                         [
                           -1.155,
-                          8.601
-                        ]
+                          8.601,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 12",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ind": 13,
@@ -1288,52 +1288,52 @@ export const winteryMixDay = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           -3.103,
-                          11.773
+                          11.773,
                         ],
                         [
                           -4.877,
-                          10
+                          10,
                         ],
                         [
                           -6.65,
-                          11.773
-                        ]
+                          11.773,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 13",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ind": 14,
@@ -1345,52 +1345,52 @@ export const winteryMixDay = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           -3.103,
-                          -2.02
+                          -2.02,
                         ],
                         [
                           -4.877,
-                          -0.247
+                          -0.247,
                         ],
                         [
                           -6.65,
-                          -2.02
-                        ]
+                          -2.02,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 14",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ind": 15,
@@ -1402,40 +1402,40 @@ export const winteryMixDay = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           -4.865,
-                          -0.365
+                          -0.365,
                         ],
                         [
                           -4.865,
-                          2.929
-                        ]
+                          2.929,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 15",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ind": 16,
@@ -1447,40 +1447,40 @@ export const winteryMixDay = {
                       "i": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "o": [
                         [
                           0,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          0
-                        ]
+                          0,
+                        ],
                       ],
                       "v": [
                         [
                           -4.866,
-                          6.844
+                          6.844,
                         ],
                         [
                           -4.866,
-                          10.138
-                        ]
+                          10.138,
+                        ],
                       ],
-                      "c": false
+                      "c": false,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 16",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ty": "st",
@@ -1490,26 +1490,26 @@ export const winteryMixDay = {
                       0.258823529412,
                       0.521568627451,
                       0.956862745098,
-                      1
+                      1,
                     ],
-                    "ix": 3
+                    "ix": 3,
                   },
                   "o": {
                     "a": 0,
                     "k": 100,
-                    "ix": 4
+                    "ix": 4,
                   },
                   "w": {
                     "a": 0,
                     "k": 0.909,
-                    "ix": 5
+                    "ix": 5,
                   },
                   "lc": 1,
                   "lj": 1,
                   "ml": 10,
                   "nm": "Stroke 1",
                   "mn": "ADBE Vector Graphic - Stroke",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ty": "tr",
@@ -1517,55 +1517,55 @@ export const winteryMixDay = {
                     "a": 0,
                     "k": [
                       12.908,
-                      2.155
+                      2.155,
                     ],
-                    "ix": 2
+                    "ix": 2,
                   },
                   "a": {
                     "a": 0,
                     "k": [
                       0,
-                      0
+                      0,
                     ],
-                    "ix": 1
+                    "ix": 1,
                   },
                   "s": {
                     "a": 0,
                     "k": [
                       100,
-                      100
+                      100,
                     ],
-                    "ix": 3
+                    "ix": 3,
                   },
                   "r": {
                     "a": 0,
                     "k": 0,
-                    "ix": 6
+                    "ix": 6,
                   },
                   "o": {
                     "a": 0,
                     "k": 100,
-                    "ix": 7
+                    "ix": 7,
                   },
                   "sk": {
                     "a": 0,
                     "k": 0,
-                    "ix": 4
+                    "ix": 4,
                   },
                   "sa": {
                     "a": 0,
                     "k": 0,
-                    "ix": 5
+                    "ix": 5,
                   },
-                  "nm": "Transform"
-                }
+                  "nm": "Transform",
+                },
               ],
               "nm": "Group 2",
               "np": 18,
               "cix": 2,
               "ix": 1,
               "mn": "ADBE Vector Group",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "gr",
@@ -1580,64 +1580,64 @@ export const winteryMixDay = {
                       "i": [
                         [
                           -1.039,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          1.039
+                          1.039,
                         ],
                         [
                           1.039,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          -1.039
-                        ]
+                          -1.039,
+                        ],
                       ],
                       "o": [
                         [
                           1.039,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          -1.039
+                          -1.039,
                         ],
                         [
                           -1.039,
-                          0
+                          0,
                         ],
                         [
                           0,
-                          1.039
-                        ]
+                          1.039,
+                        ],
                       ],
                       "v": [
                         [
                           8.031,
-                          8.912
+                          8.912,
                         ],
                         [
                           9.912,
-                          7.031
+                          7.031,
                         ],
                         [
                           8.031,
-                          5.15
+                          5.15,
                         ],
                         [
                           6.15,
-                          7.031
-                        ]
+                          7.031,
+                        ],
                       ],
-                      "c": true
+                      "c": true,
                     },
-                    "ix": 2
+                    "ix": 2,
                   },
                   "nm": "Path 18",
                   "mn": "ADBE Vector Shape - Group",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ty": "fl",
@@ -1647,19 +1647,19 @@ export const winteryMixDay = {
                       1,
                       1,
                       1,
-                      1
+                      1,
                     ],
-                    "ix": 4
+                    "ix": 4,
                   },
                   "o": {
                     "a": 0,
                     "k": 100,
-                    "ix": 5
+                    "ix": 5,
                   },
                   "r": 1,
                   "nm": "Fill 1",
                   "mn": "ADBE Vector Graphic - Fill",
-                  "hd": false
+                  "hd": false,
                 },
                 {
                   "ty": "tr",
@@ -1667,64 +1667,64 @@ export const winteryMixDay = {
                     "a": 0,
                     "k": [
                       0,
-                      0
+                      0,
                     ],
-                    "ix": 2
+                    "ix": 2,
                   },
                   "a": {
                     "a": 0,
                     "k": [
                       0,
-                      0
+                      0,
                     ],
-                    "ix": 1
+                    "ix": 1,
                   },
                   "s": {
                     "a": 0,
                     "k": [
                       100,
-                      100
+                      100,
                     ],
-                    "ix": 3
+                    "ix": 3,
                   },
                   "r": {
                     "a": 0,
                     "k": 0,
-                    "ix": 6
+                    "ix": 6,
                   },
                   "o": {
                     "a": 0,
                     "k": 100,
-                    "ix": 7
+                    "ix": 7,
                   },
                   "sk": {
                     "a": 0,
                     "k": 0,
-                    "ix": 4
+                    "ix": 4,
                   },
                   "sa": {
                     "a": 0,
                     "k": 0,
-                    "ix": 5
+                    "ix": 5,
                   },
-                  "nm": "Transform"
-                }
+                  "nm": "Transform",
+                },
               ],
               "nm": "Group 1",
               "np": 2,
               "cix": 2,
               "ix": 2,
               "mn": "ADBE Vector Group",
-              "hd": false
-            }
+              "hd": false,
+            },
           ],
           "ip": 0,
           "op": 848,
           "st": 0,
-          "bm": 0
-        }
-      ]
-    }
+          "bm": 0,
+        },
+      ],
+    },
   ],
   "layers": [
     {
@@ -1741,25 +1741,25 @@ export const winteryMixDay = {
             {
               "i": {
                 "x": [0.833],
-                "y": [0.833]
+                "y": [0.833],
               },
               "o": {
                 "x": [0.167],
-                "y": [0.167]
+                "y": [0.167],
               },
               "n": ["0p833_0p833_0p167_0p167"],
               "t": 45,
               "s": [0],
-              "e": [100]
+              "e": [100],
             },
-            {"t": 91}
+            {"t": 91},
           ],
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 1,
@@ -1767,57 +1767,57 @@ export const winteryMixDay = {
             {
               "i": {
                 "x": 0.36,
-                "y": 1
+                "y": 1,
               },
               "o": {
                 "x": 0.22,
-                "y": 1
+                "y": 1,
               },
               "n": "0p36_1_0p22_1",
               "t": 45,
               "s": [
                 1117,
                 505,
-                0
+                0,
               ],
               "e": [
                 1117,
                 645,
-                0
+                0,
               ],
               "to": [
                 0,
                 23.3333339691162,
-                0
+                0,
               ],
               "ti": [
                 0,
                 -23.3333339691162,
-                0
-              ]
+                0,
+              ],
             },
-            {"t": 91}
+            {"t": 91},
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             105,
             105,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             90,
             90,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "w": 210,
@@ -1825,7 +1825,7 @@ export const winteryMixDay = {
       "ip": 16,
       "op": 3616,
       "st": 16,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -1841,25 +1841,25 @@ export const winteryMixDay = {
             {
               "i": {
                 "x": [0.833],
-                "y": [0.833]
+                "y": [0.833],
               },
               "o": {
                 "x": [0.167],
-                "y": [0.167]
+                "y": [0.167],
               },
               "n": ["0p833_0p833_0p167_0p167"],
               "t": 35,
               "s": [0],
-              "e": [100]
+              "e": [100],
             },
-            {"t": 81}
+            {"t": 81},
           ],
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 1,
@@ -1867,57 +1867,57 @@ export const winteryMixDay = {
             {
               "i": {
                 "x": 0.36,
-                "y": 1
+                "y": 1,
               },
               "o": {
                 "x": 0.22,
-                "y": 1
+                "y": 1,
               },
               "n": "0p36_1_0p22_1",
               "t": 35,
               "s": [
                 861,
                 505,
-                0
+                0,
               ],
               "e": [
                 861,
                 645,
-                0
+                0,
               ],
               "to": [
                 0,
                 23.3333339691162,
-                0
+                0,
               ],
               "ti": [
                 0,
                 -23.3333339691162,
-                0
-              ]
+                0,
+              ],
             },
-            {"t": 81}
+            {"t": 81},
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             105,
             105,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             90,
             90,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "w": 210,
@@ -1925,7 +1925,7 @@ export const winteryMixDay = {
       "ip": 6,
       "op": 3606,
       "st": 6,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -1941,20 +1941,20 @@ export const winteryMixDay = {
             {
               "i": {
                 "x": [0.833],
-                "y": [0.833]
+                "y": [0.833],
               },
               "o": {
                 "x": [0.167],
-                "y": [0.167]
+                "y": [0.167],
               },
               "n": ["0p833_0p833_0p167_0p167"],
               "t": 58,
               "s": [0],
-              "e": [100]
+              "e": [100],
             },
-            {"t": 104}
+            {"t": 104},
           ],
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 1,
@@ -1962,20 +1962,20 @@ export const winteryMixDay = {
             {
               "i": {
                 "x": [0.36],
-                "y": [1]
+                "y": [1],
               },
               "o": {
                 "x": [0.22],
-                "y": [1]
+                "y": [1],
               },
               "n": ["0p36_1_0p22_1"],
               "t": 58,
               "s": [-45],
-              "e": [0]
+              "e": [0],
             },
-            {"t": 104}
+            {"t": 104},
           ],
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 1,
@@ -1983,57 +1983,57 @@ export const winteryMixDay = {
             {
               "i": {
                 "x": 0.36,
-                "y": 1
+                "y": 1,
               },
               "o": {
                 "x": 0.22,
-                "y": 1
+                "y": 1,
               },
               "n": "0p36_1_0p22_1",
               "t": 58,
               "s": [
                 989,
                 713,
-                0
+                0,
               ],
               "e": [
                 989,
                 981,
-                0
+                0,
               ],
               "to": [
                 0,
                 44.6666679382324,
-                0
+                0,
               ],
               "ti": [
                 0,
                 -44.6666679382324,
-                0
-              ]
+                0,
+              ],
             },
-            {"t": 104}
+            {"t": 104},
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             12,
             12,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             2874,
             2874,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "w": 24,
@@ -2041,7 +2041,7 @@ export const winteryMixDay = {
       "ip": 58,
       "op": 906,
       "st": 58,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -2053,12 +2053,12 @@ export const winteryMixDay = {
         "o": {
           "a": 0,
           "k": 0,
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 1,
@@ -2066,47 +2066,47 @@ export const winteryMixDay = {
             {
               "i": {
                 "x": 0.36,
-                "y": 1
+                "y": 1,
               },
               "o": {
                 "x": 0.22,
-                "y": 1
+                "y": 1,
               },
               "n": "0p36_1_0p22_1",
               "t": 1,
               "s": [
                 422,
                 750,
-                0
+                0,
               ],
               "e": [
                 750,
                 750,
-                0
+                0,
               ],
               "to": [
                 54.6666679382324,
                 0,
-                0
+                0,
               ],
               "ti": [
                 -54.6666679382324,
                 0,
-                0
-              ]
+                0,
+              ],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             0,
             0,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 1,
@@ -2116,53 +2116,53 @@ export const winteryMixDay = {
                 "x": [
                   0.36,
                   0.36,
-                  0.36
+                  0.36,
                 ],
                 "y": [
                   1,
                   1,
-                  1
-                ]
+                  1,
+                ],
               },
               "o": {
                 "x": [
                   0.22,
                   0.22,
-                  0.22
+                  0.22,
                 ],
                 "y": [
                   0,
                   0,
-                  0
-                ]
+                  0,
+                ],
               },
               "n": [
                 "0p36_1_0p22_0",
                 "0p36_1_0p22_0",
-                "0p36_1_0p22_0"
+                "0p36_1_0p22_0",
               ],
               "t": 1,
               "s": [
                 100,
                 100,
-                100
+                100,
               ],
               "e": [
                 100,
                 100,
-                100
-              ]
+                100,
+              ],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "ip": -33,
       "op": 849,
       "st": -2,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -2178,53 +2178,53 @@ export const winteryMixDay = {
             {
               "i": {
                 "x": [0.518],
-                "y": [0.936]
+                "y": [0.936],
               },
               "o": {
                 "x": [0.294],
-                "y": [0.52]
+                "y": [0.52],
               },
               "n": ["0p518_0p936_0p294_0p52"],
               "t": 1,
               "s": [0],
-              "e": [100]
+              "e": [100],
             },
-            {"t": 37}
+            {"t": 37},
           ],
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 0,
           "k": [
             -250,
             -500,
-            0
+            0,
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             0,
             0,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             100,
             100,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "shapes": [
@@ -2238,21 +2238,21 @@ export const winteryMixDay = {
                 "a": 0,
                 "k": [
                   24,
-                  24
+                  24,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "p": {
                 "a": 0,
                 "k": [
                   12,
-                  12
+                  12,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "nm": "Ellipse Path 1",
               "mn": "ADBE Vector Shape - Ellipse",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "fl",
@@ -2262,26 +2262,26 @@ export const winteryMixDay = {
                   0.945097982883,
                   0.952941000462,
                   0.956862986088,
-                  1
+                  1,
                 ],
-                "ix": 4
+                "ix": 4,
               },
               "o": {
                 "a": 0,
                 "k": 80,
-                "ix": 5
+                "ix": 5,
               },
               "r": 2,
               "nm": "Fill 1",
               "mn": "ADBE Vector Graphic - Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "gf",
               "o": {
                 "a": 0,
                 "k": 50,
-                "ix": 10
+                "ix": 10,
               },
               "r": 2,
               "g": {
@@ -2306,31 +2306,31 @@ export const winteryMixDay = {
                     0.5,
                     0.1,
                     1,
-                    0.2
+                    0.2,
                   ],
-                  "ix": 9
-                }
+                  "ix": 9,
+                },
               },
               "s": {
                 "a": 0,
                 "k": [
                   11.385,
-                  13.566
+                  13.566,
                 ],
-                "ix": 5
+                "ix": 5,
               },
               "e": {
                 "a": 0,
                 "k": [
                   23.401,
-                  24.13
+                  24.13,
                 ],
-                "ix": 6
+                "ix": 6,
               },
               "t": 1,
               "nm": "Gradient Fill 1",
               "mn": "ADBE Vector Graphic - G-Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "tr",
@@ -2338,61 +2338,61 @@ export const winteryMixDay = {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "a": {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 1
+                "ix": 1,
               },
               "s": {
                 "a": 0,
                 "k": [
                   3125,
-                  3125
+                  3125,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "r": {
                 "a": 0,
                 "k": 0,
-                "ix": 6
+                "ix": 6,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 7
+                "ix": 7,
               },
               "sk": {
                 "a": 0,
                 "k": 0,
-                "ix": 4
+                "ix": 4,
               },
               "sa": {
                 "a": 0,
                 "k": 0,
-                "ix": 5
+                "ix": 5,
               },
-              "nm": "Transform"
-            }
+              "nm": "Transform",
+            },
           ],
           "nm": "Oval",
           "np": 3,
           "cix": 2,
           "ix": 1,
           "mn": "ADBE Vector Group",
-          "hd": false
-        }
+          "hd": false,
+        },
       ],
       "ip": -58,
       "op": 842,
       "st": -6,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -2407,25 +2407,25 @@ export const winteryMixDay = {
             {
               "i": {
                 "x": [0.36],
-                "y": [1]
+                "y": [1],
               },
               "o": {
                 "x": [0.22],
-                "y": [99]
+                "y": [99],
               },
               "n": ["0p36_1_0p22_99"],
               "t": 1,
               "s": [100],
-              "e": [100]
+              "e": [100],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 1,
@@ -2433,57 +2433,57 @@ export const winteryMixDay = {
             {
               "i": {
                 "x": 0.36,
-                "y": 1
+                "y": 1,
               },
               "o": {
                 "x": 0.22,
-                "y": 1
+                "y": 1,
               },
               "n": "0p36_1_0p22_1",
               "t": 1,
               "s": [
                 487.75,
                 237,
-                0
+                0,
               ],
               "e": [
                 232.75,
                 237,
-                0
+                0,
               ],
               "to": [
                 -42.5,
                 0,
-                0
+                0,
               ],
               "ti": [
                 42.5,
                 0,
-                0
-              ]
+                0,
+              ],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             -496,
             -494,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             100,
             100,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "shapes": [
@@ -2497,21 +2497,21 @@ export const winteryMixDay = {
                 "a": 0,
                 "k": [
                   365.547,
-                  365.547
+                  365.547,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "p": {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "nm": "Ellipse Path 1",
               "mn": "ADBE Vector Shape - Ellipse",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "tr",
@@ -2519,55 +2519,55 @@ export const winteryMixDay = {
                 "a": 0,
                 "k": [
                   -499.359,
-                  -499.43
+                  -499.43,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "a": {
                 "a": 0,
                 "k": [
                   -209.552,
-                  -208.357
+                  -208.357,
                 ],
-                "ix": 1
+                "ix": 1,
               },
               "s": {
                 "a": 0,
                 "k": [
                   131.101,
-                  128.967
+                  128.967,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "r": {
                 "a": 0,
                 "k": 0,
-                "ix": 6
+                "ix": 6,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 7
+                "ix": 7,
               },
               "sk": {
                 "a": 0,
                 "k": 0,
-                "ix": 4
+                "ix": 4,
               },
               "sa": {
                 "a": 0,
                 "k": 0,
-                "ix": 5
+                "ix": 5,
               },
-              "nm": "Transform"
-            }
+              "nm": "Transform",
+            },
           ],
           "nm": "Ellipse 1",
           "np": 1,
           "cix": 2,
           "ix": 1,
           "mn": "ADBE Vector Group",
-          "hd": false
+          "hd": false,
         },
         {
           "ty": "st",
@@ -2577,14 +2577,14 @@ export const winteryMixDay = {
               0.98431372549,
               0.737254901961,
               0.01568627451,
-              1
+              1,
             ],
-            "ix": 3
+            "ix": 3,
           },
           "o": {
             "a": 0,
             "k": 100,
-            "ix": 4
+            "ix": 4,
           },
           "w": {
             "a": 1,
@@ -2592,33 +2592,33 @@ export const winteryMixDay = {
               {
                 "i": {
                   "x": [0.36],
-                  "y": [1]
+                  "y": [1],
                 },
                 "o": {
                   "x": [0.22],
-                  "y": [1]
+                  "y": [1],
                 },
                 "n": ["0p36_1_0p22_1"],
                 "t": 1,
                 "s": [1],
-                "e": [30]
+                "e": [30],
               },
-              {"t": 75}
+              {"t": 75},
             ],
-            "ix": 5
+            "ix": 5,
           },
           "lc": 1,
           "lj": 1,
           "ml": 4,
           "nm": "Stroke 1",
           "mn": "ADBE Vector Graphic - Stroke",
-          "hd": false
-        }
+          "hd": false,
+        },
       ],
       "ip": 1,
       "op": 901,
       "st": -1,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -2632,40 +2632,40 @@ export const winteryMixDay = {
         "o": {
           "a": 0,
           "k": 100,
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 0,
           "k": [
             -250,
             -500,
-            0
+            0,
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             0,
             0,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             100,
             100,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "shapes": [
@@ -2679,21 +2679,21 @@ export const winteryMixDay = {
                 "a": 0,
                 "k": [
                   24,
-                  24
+                  24,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "p": {
                 "a": 0,
                 "k": [
                   12,
-                  12
+                  12,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "nm": "Ellipse Path 1",
               "mn": "ADBE Vector Shape - Ellipse",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "fl",
@@ -2703,26 +2703,26 @@ export const winteryMixDay = {
                   0.945097982883,
                   0.952941000462,
                   0.956862986088,
-                  1
+                  1,
                 ],
-                "ix": 4
+                "ix": 4,
               },
               "o": {
                 "a": 0,
                 "k": 80,
-                "ix": 5
+                "ix": 5,
               },
               "r": 2,
               "nm": "Fill 1",
               "mn": "ADBE Vector Graphic - Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "gf",
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 10
+                "ix": 10,
               },
               "r": 2,
               "g": {
@@ -2747,31 +2747,31 @@ export const winteryMixDay = {
                     0.5,
                     0.1,
                     1,
-                    0.2
+                    0.2,
                   ],
-                  "ix": 9
-                }
+                  "ix": 9,
+                },
               },
               "s": {
                 "a": 0,
                 "k": [
                   11.385,
-                  13.566
+                  13.566,
                 ],
-                "ix": 5
+                "ix": 5,
               },
               "e": {
                 "a": 0,
                 "k": [
                   23.401,
-                  24.13
+                  24.13,
                 ],
-                "ix": 6
+                "ix": 6,
               },
               "t": 1,
               "nm": "Gradient Fill 1",
               "mn": "ADBE Vector Graphic - G-Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "tr",
@@ -2779,61 +2779,61 @@ export const winteryMixDay = {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "a": {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 1
+                "ix": 1,
               },
               "s": {
                 "a": 0,
                 "k": [
                   3125,
-                  3125
+                  3125,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "r": {
                 "a": 0,
                 "k": 0,
-                "ix": 6
+                "ix": 6,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 7
+                "ix": 7,
               },
               "sk": {
                 "a": 0,
                 "k": 0,
-                "ix": 4
+                "ix": 4,
               },
               "sa": {
                 "a": 0,
                 "k": 0,
-                "ix": 5
+                "ix": 5,
               },
-              "nm": "Transform"
-            }
+              "nm": "Transform",
+            },
           ],
           "nm": "Oval",
           "np": 3,
           "cix": 2,
           "ix": 1,
           "mn": "ADBE Vector Group",
-          "hd": false
-        }
+          "hd": false,
+        },
       ],
       "ip": -58,
       "op": 842,
       "st": -6,
-      "bm": 0
+      "bm": 0,
     },
     {
       "ddd": 0,
@@ -2850,53 +2850,53 @@ export const winteryMixDay = {
             {
               "i": {
                 "x": [0.36],
-                "y": [1]
+                "y": [1],
               },
               "o": {
                 "x": [0.22],
-                "y": [1]
+                "y": [1],
               },
               "n": ["0p36_1_0p22_1"],
               "t": 1,
               "s": [0],
-              "e": [100]
+              "e": [100],
             },
-            {"t": 75}
+            {"t": 75},
           ],
-          "ix": 11
+          "ix": 11,
         },
         "r": {
           "a": 0,
           "k": 0,
-          "ix": 10
+          "ix": 10,
         },
         "p": {
           "a": 0,
           "k": [
             -84,
             -62.5,
-            0
+            0,
           ],
-          "ix": 2
+          "ix": 2,
         },
         "a": {
           "a": 0,
           "k": [
             416,
             0,
-            0
+            0,
           ],
-          "ix": 1
+          "ix": 1,
         },
         "s": {
           "a": 0,
           "k": [
             100,
             100,
-            100
+            100,
           ],
-          "ix": 6
-        }
+          "ix": 6,
+        },
       },
       "ao": 0,
       "shapes": [
@@ -2913,52 +2913,52 @@ export const winteryMixDay = {
                   "i": [
                     [
                       5.52,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      -5.52
+                      -5.52,
                     ],
                     [
                       0.01,
-                      -0.02
-                    ]
+                      -0.02,
+                    ],
                   ],
                   "o": [
                     [
                       -5.52,
-                      0
+                      0,
                     ],
                     [
                       0.01,
-                      -0.02
+                      -0.02,
                     ],
                     [
                       0,
-                      -5.52
-                    ]
+                      -5.52,
+                    ],
                   ],
                   "v": [
                     [
                       10,
-                      0
+                      0,
                     ],
                     [
                       0,
-                      10
+                      10,
                     ],
                     [
                       20,
-                      10
-                    ]
+                      10,
+                    ],
                   ],
-                  "c": true
+                  "c": true,
                 },
-                "ix": 2
+                "ix": 2,
               },
               "nm": "Path 1",
               "mn": "ADBE Vector Shape - Group",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "fl",
@@ -2968,19 +2968,19 @@ export const winteryMixDay = {
                   0.945097982883,
                   0.952941000462,
                   0.956862986088,
-                  1
+                  1,
                 ],
-                "ix": 4
+                "ix": 4,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 5
+                "ix": 5,
               },
               "r": 1,
               "nm": "Fill 1",
               "mn": "ADBE Vector Graphic - Fill",
-              "hd": false
+              "hd": false,
             },
             {
               "ty": "tr",
@@ -2988,62 +2988,62 @@ export const winteryMixDay = {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 2
+                "ix": 2,
               },
               "a": {
                 "a": 0,
                 "k": [
                   0,
-                  0
+                  0,
                 ],
-                "ix": 1
+                "ix": 1,
               },
               "s": {
                 "a": 0,
                 "k": [
                   3125,
-                  3125
+                  3125,
                 ],
-                "ix": 3
+                "ix": 3,
               },
               "r": {
                 "a": 0,
                 "k": 0,
-                "ix": 6
+                "ix": 6,
               },
               "o": {
                 "a": 0,
                 "k": 100,
-                "ix": 7
+                "ix": 7,
               },
               "sk": {
                 "a": 0,
                 "k": 0,
-                "ix": 4
+                "ix": 4,
               },
               "sa": {
                 "a": 0,
                 "k": 0,
-                "ix": 5
+                "ix": 5,
               },
-              "nm": "Transform"
-            }
+              "nm": "Transform",
+            },
           ],
           "nm": "Group 1",
           "np": 2,
           "cix": 2,
           "ix": 1,
           "mn": "ADBE Vector Group",
-          "hd": false
-        }
+          "hd": false,
+        },
       ],
       "ip": -52,
       "op": 848,
       "st": -70,
-      "bm": 0
-    }
+      "bm": 0,
+    },
   ],
-  "markers": []
+  "markers": [],
 };
